% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_bbox.R
\name{adjust_bbox}
\alias{adjust_bbox}
\title{Get bounding box buffered and adjusted to aspect ratio}
\usage{
adjust_bbox(
  area = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crs = NULL
)
}
\arguments{
\item{area}{\code{sf} object to buffer and/or adjust.}

\item{bbox}{\code{bbox} object to buffer and/or adjust. If an area is provided, any bbox is ignored.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crs}{Coordinate reference system of bounding box to return}
}
\value{
Class \code{bbox} object
}
\description{
Get bounding box buffered and adjusted to match aspect ratio
}
\details{
Takes an area as an sf object or a bounding box and returns a bounding box
that matches the provided aspect ratio and contains the area or bounding box
provided.

Common aspect ratios include "1:1" (1), "4:6" (0.666), "8.5:11", "16:9"
(1.777). The asp parameter supports both numeric values and character
strings with ratios matching the format of "width:height".
}
