% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_bbox_asp.R
\name{adjust_bbox_asp}
\alias{adjust_bbox_asp}
\title{Get bounding box adjusted to match aspect ratio}
\usage{
adjust_bbox_asp(area = NULL, bbox = NULL, asp = NULL)
}
\arguments{
\item{area}{\code{sf} object}

\item{bbox}{\code{bbox} object to adjust}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}
}
\value{
\code{bbox} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because the functionality has been incorporated
into the improved \code{\link[sfext:st_bbox_ext]{sfext::st_bbox_asp()}} function which uses a similar set
of parameters.

Get bbox from sf or bbox object adjusted to match an aspect ratio

Takes an area as an  \code{sf} or \code{bbox} object and returns a bounding
box that matches the aspect ratio provided to \code{asp} and contains the
area or bounding box provided. Common aspect ratios include "1:1" (1), "4:6"
(0.666), "8.5:11", "16:9" (1.777). The asp parameter supports both numeric
values and character strings with ratios matching the format of
"width:height".
}
\keyword{internal}
