% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{balt_tbl_labs}
\alias{balt_tbl_labs}
\title{Baltimore Data Table Labels}
\format{
A data frame with 9 rows and 7 variables:
\describe{
\item{\code{fn_name}}{character Function name}
\item{\code{table}}{character Table name}
\item{\code{col}}{character Column name}
\item{\code{label}}{character Column label}
\item{\code{definition}}{logical Column variable definition}
\item{\code{source}}{logical Column variable data source}
\item{\code{fmt}}{character Column data format}
}
}
\source{
\url{https://docs.google.com/spreadsheets/d/1FXEJlhccnhoQmSO2WydBidXIw-f2lpomURDGy9KBgJw/edit?usp=sharing}
}
\usage{
balt_tbl_labs
}
\description{
A data.frame with labels to use with tables created using mapbaltimore data.
The Housing Market Typology 2017 data is the only set of labels included and
the preset table functions are not yet implemented.
}
\keyword{datasets}
