% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_bbox}
\alias{baltimore_bbox}
\title{Baltimore City WGS84 Bounding Box}
\format{
A \code{bbox} class object.
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_bbox
}
\description{
A generalized boundary for Baltimore City, Maryland (\code{baltimore_city})
converted to a bounding box object using a EPSG:4326 coordinate reference
system.
}
\keyword{datasets}
