% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_blocks}
\alias{baltimore_blocks}
\title{U.S. Census Blocks (2010 Decennial) in Baltimore City}
\format{
A data frame with 13,598 rows and 9 variables:
\describe{
  \item{tractce10}{...}
  \item{blkgrpce10}{...}
  \item{geoid10}{...}
  \item{name10}{...}
  \item{aland10}{...}
  \item{awater10}{...}
  \item{intptlat10}{...}
  \item{intptlon10}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{...}
}
\usage{
baltimore_blocks
}
\description{
...
}
\keyword{datasets}
