% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_census_xwalk}
\alias{baltimore_census_xwalk}
\title{Crosswalk for Baltimore areas and Census geography}
\format{
A data frame with 1024 rows and 8 variables:
\describe{
\item{\code{geography}}{Geography/area type}
\item{\code{GEOID}}{2020 Census GeoID}
\item{\code{TRACTCE20}}{2020 Census Tract ID}
\item{\code{name}}{Area name}
\item{\code{POP20}}{Population in area and tract}
\item{\code{perc_POP20}}{Percent of population in area and tract}
\item{\code{HOUSING20}}{Households/occupied housing units in area and tract}
\item{\code{perc_HOUSING20}}{Percent of households in area and tract}
}
}
\usage{
baltimore_census_xwalk
}
\description{
A pre-built crosswalk data frame that can be filtered by geography and then
used with the \code{\link[getACS:make_area_xwalk]{getACS::use_area_xwalk()}} function. Crosswalk uses the 2010
City Council district boundaries, 2010 and 2020 neighborhood (neighborhood
statistical area) boundaries, and current city planning districts.
}
\details{
Created using the \code{\link[getACS:make_area_xwalk]{getACS::make_area_xwalk()}} data.
}
\keyword{datasets}
