% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_gis_index}
\alias{baltimore_gis_index}
\title{Baltimore ArcGIS Server index data}
\format{
A data frame with 1020 rows and 15 variables:
\describe{
\item{\code{name}}{character COLUMN_DESCRIPTION}
\item{\code{nm}}{character COLUMN_DESCRIPTION}
\item{\code{type}}{character COLUMN_DESCRIPTION}
\item{\code{url}}{character COLUMN_DESCRIPTION}
\item{\code{urlType}}{character COLUMN_DESCRIPTION}
\item{\code{folderPath}}{character COLUMN_DESCRIPTION}
\item{\code{serviceName}}{character COLUMN_DESCRIPTION}
\item{\code{serviceType}}{character COLUMN_DESCRIPTION}
\item{\code{id}}{integer COLUMN_DESCRIPTION}
\item{\code{parentLayerId}}{integer COLUMN_DESCRIPTION}
\item{\code{defaultVisibility}}{logical COLUMN_DESCRIPTION}
\item{\code{minScale}}{double COLUMN_DESCRIPTION}
\item{\code{maxScale}}{integer COLUMN_DESCRIPTION}
\item{\code{geometryType}}{character COLUMN_DESCRIPTION}
\item{\code{subLayerIds}}{list COLUMN_DESCRIPTION}
}
}
\usage{
baltimore_gis_index
}
\description{
A dataframe indexing the layers, services, and folders on two ArcGIS Servers
maintained by the Baltimore City Mayor's Office of Information Technology
(MOIT) Enterprise GIS (EGIS) program. Use by the \code{\link[=get_baltimore_data]{get_baltimore_data()}}
function.
}
\details{
DETAILS
}
\keyword{datasets}
