% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_mihp}
\alias{baltimore_mihp}
\title{Maryland Inventory of Historic Properties in Baltimore City}
\format{
A data frame with 5,203 rows and 14 variables:
\describe{
  \item{num_polys}{...}
  \item{mihp_id}{...}
  \item{property_id}{...}
  \item{mihp_num}{...}
  \item{name}{...}
  \item{alternate_name}{...}
  \item{full_address}{...}
  \item{town}{...}
  \item{county}{...}
  \item{pdflink}{...}
  \item{xcoord}{...}
  \item{ycoord}{...}
  \item{do_erecord}{...}
  \item{geoms}{...}
  ...
}
}
\source{
\url{...}
}
\usage{
baltimore_mihp
}
\description{
...
}
\keyword{datasets}
