% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_msa_water}
\alias{baltimore_msa_water}
\title{Baltimore Metropolitan Statistical Area (MSA) Water Polygons}
\format{
A data frame with 3,491 rows and 9 variables:
\describe{
\item{\code{ansicode}}{American National Standards Institute codes (ANSI codes)}
\item{\code{hydroid}}{Unique key for hydrographic features}
\item{\code{fullname}}{Full name}
\item{\code{mtfcc}}{MAF/TIGER Feature Class Code}
\item{\code{aland}}{land area (square meters)}
\item{\code{awater}}{water area (square meters)}
\item{\code{intptlat}}{latitude of the internal point}
\item{\code{intptlon}}{longitude of the internal point}
\item{\code{geometry}}{POLYGON geometry}
}
}
\usage{
baltimore_msa_water
}
\description{
Downloaded using tigris package.
}
\keyword{datasets}
