% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_pumas}
\alias{baltimore_pumas}
\title{Baltimore PUMAS (Public Use Microdata Areas) - 2010}
\format{
A data frame with 5 rows and 8 variables:
\describe{
\item{\code{pumace10}}{PUMA code}
\item{\code{geoid10}}{GeoID}
\item{\code{namelsad10}}{name and the translated legal/statistical area description code for census tract}
\item{\code{aland10}}{land area (square meters)}
\item{\code{awater10}}{water area (square meters)}
\item{\code{intptlat10}}{latitude of the internal point}
\item{\code{intptlon10}}{longitude of the internal point}
\item{\code{geometry}}{Polygon with PUMA boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_pumas
}
\description{
The U.S. Census Bureau explains that "Public Use Microdata Areas
(PUMAs) are non-overlapping, statistical geographic areas that partition
each state or equivalent entity into geographic areas containing no fewer
than 100,000 people each... The Census Bureau defines PUMAs for the
tabulation and dissemination of decennial census and American Community
Survey (ACS) Public Use Microdata Sample (PUMS) data."
}
\keyword{datasets}
