% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_water}
\alias{baltimore_water}
\title{Baltimore City Water}
\format{
A data frame with 468 rows and 6 variables:
\describe{
\item{\code{name}}{Water feature name, if available}
\item{\code{type}}{Water type}
\item{\code{subtype}}{Water subtype}
\item{\code{symbol}}{Symbol}
\item{\code{water}}{Water indicator}
\item{\code{acres}}{Feature area in acres}
\item{\code{geometry}}{MULTIPOLYGON geometry}
}
}
\source{
\url{https://dotgis.baltimorecity.gov/arcgis/rest/services/DOT_Map_Services/DOT_Basemap/MapServer/7}
}
\usage{
baltimore_water
}
\description{
Detailed MULTIPOLYGON data for area of streams, lakes, and other water bodies
in Baltimore City.
}
\keyword{datasets}
