% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_area.R
\name{buffer_area}
\alias{buffer_area}
\title{Get buffered area}
\usage{
buffer_area(area, dist = NULL, diag_ratio = NULL)
}
\arguments{
\item{area}{sf object.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because the functionality has been incorporated
into the improved \code{\link[sfext:st_buffer_ext]{sfext::st_buffer_ext()}} function which uses a similar set
of parameters.

Return an sf object of an area with a buffer applied to it. If no buffer
distance is provided, a default buffer is calculated of one-eighth the
diagonal distance of the bounding box (corner to corner) for the area. The
metadata for the provided area remains the same.
}
\keyword{internal}
