% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buildings_21stc}
\alias{buildings_21stc}
\title{Baltimore 21st Century Schools}
\format{
A data frame with 28 rows and 20 variables:
\describe{
\item{\code{bldg_name}}{Building name}
\item{\code{name}}{Name (identical to build name)}
\item{\code{bldg_name_short}}{Short building name}
\item{\code{project_type}}{Project type}
\item{\code{project_url}}{Project URL}
\item{\code{building_occupied}}{Building occupied year/season (or scheduled occupation date)}
\item{\code{inspire_plan}}{INSPIRE Plan}
\item{\code{inspire_plan_url}}{INSPIRE Plan URL}
\item{\code{school_names}}{School/program names}
\item{\code{school_names_short}}{Short school/program names}
\item{\code{school_numbers}}{School numbers}
\item{\code{grade_bands}}{Schools grade bands}
\item{\code{grades_served}}{Grades served}
\item{\code{address}}{Street address}
\item{\code{city}}{City}
\item{\code{state}}{State}
\item{\code{zip}}{Zip code}
\item{\code{lon}}{Longitude (EPSG 4326)}
\item{\code{lat}}{Latitude (EPSG 4326)}
\item{\code{geometry}}{POINT geometry for building locations}
}
}
\usage{
buildings_21stc
}
\description{
Buildings constructed or renovated through the \href{https://baltimore21stcenturyschools.org/}{21st Century Schools Program}. See \code{schools_21stc} for
school-level information.
}
\keyword{datasets}
