% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{cache_mapbaltimore_data}
\alias{cache_mapbaltimore_data}
\title{Cache data for mapbaltimore package}
\usage{
cache_mapbaltimore_data(crs = 2804, cache_dir_path = NULL, overwrite = FALSE)
}
\arguments{
\item{crs}{Coordinate reference system. Default 2804.}

\item{cache_dir_path}{Cache directory path. Default is NULL which sets path to `rappdirs::user_cache_dir("mapbaltimore")`}

\item{overwrite}{Logical. Default FALSE. If TRUE, overwrite any existing cached files that use the same filename.}
}
\description{
mapbaltimore works with several large datasets that exceed
the 100 MB file size limit for files on a GitHub repositories.
This function downloads three datasets from the city and state
ArcGIS Feature Servers using `esri2sf::esri2sf()` and save the
data as geopackage files to the cache folder location returned by
`rappdirs::user_cache_dir("mapbaltimore")`. These datasets can then
be accessed using the `cachedata` parameter of the `get_area_data` function.
}
