% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{cache_real_property}
\alias{cache_real_property}
\title{Cache real property data for Baltimore City}
\usage{
cache_real_property(slug = "real_property", crs = 2804, overwrite = FALSE)
}
\arguments{
\item{slug}{Name to use for cached file. Default to "real_property"}
}
\description{
mapbaltimore works with several large datasets that exceed
the 100 MB file size limit for files on a GitHub repositories.
This function downloads the real property data from the city
ArcGIS Feature Servers using `esri2sf::esri2sf()` and saves the
data as geopackage files to the cache folder location returned by
`rappdirs::user_cache_dir("mapbaltimore")`. This datasets can then
be accessed using the `cachedata` parameter of the `get_area_data` function.
}
