% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chap_districts}
\alias{chap_districts}
\title{CHAP Historic Districts}
\format{
A data frame with 39 rows and 7 variables:
\describe{
\item{\code{name}}{Historic district name}
\item{\code{contact_name}}{CHAP Staff contact name}
\item{\code{url}}{URL for CHAP website}
\item{\code{deed_covenant}}{Design review required under deed covenants}
\item{\code{overlaps_nr_district}}{District is also designated as or overlaps
some or entirely with a designated National Register Historic District}
\item{\code{acres}}{Acreage}
\item{\code{geometry}}{MULTIPOLYGON boundary geometry}
}
}
\usage{
chap_districts
}
\description{
Historic districts designated by the \href{https://chap.baltimorecity.gov/}{Baltimore City Commission on Historical and Architectural Preservation} (CHAP) which
is the local historic preservation office for Baltimore City, Maryland.
Updated 2023 February 10.
}
\keyword{datasets}
