% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circulator_routes}
\alias{circulator_routes}
\title{Charm City Circulator Routes}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{\code{route_name}}{Route name}
\item{\code{alt_route_name}}{Alternate route name}
\item{\code{geometry}}{MULTILINESTRING geometry with routes}
}
}
\source{
\href{https://egisdata.baltimorecity.gov/egis/rest/services/CityView/Charm_City_Circulator/MapServer/1}{Baltimore CityView - Charm City Circulator Routes}
}
\usage{
circulator_routes
}
\description{
The Baltimore City Department of Transportation describes the Charm City
Circulator (CCC) as "a fleet of 24 free shuttles that travel four routes in
the central business district of Baltimore City, Maryland." The Harbor
Connector (HC) is "an extension of the CCC and is the City’s free maritime
transit service connecting 6 piers through four vessels."
}
\keyword{datasets}
