% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_area.R
\name{clip_area}
\alias{clip_area}
\title{Clip an area to a portion of the whole area}
\usage{
clip_area(
  area,
  clip = c("outsideedge", "insideedge", "topleft", "topright", "bottomleft",
    "bottomright"),
  flip = FALSE
)
}
\arguments{
\item{area}{`sf` object to clip}

\item{clip}{Character string describing the part of the area to remove or,
if "edge", the only area to keep. Options include c("outsideedge", "insideedge", "topleft",
"topright", "bottomleft", "bottomright")}

\item{flip}{Logical. Default FALSE. If TRUE, the reverse of the select area
is removed, e.g. if clip is "topright" the "bottomleft" area is removed
instead.}
}
\value{
\code{sf} object clipped based on parameters
}
\description{
Clip based on the corner of the bounding box. Used for the labelling option
within \code{layer_area_streets}
}
