% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congressional_districts}
\alias{congressional_districts}
\title{U.S. Congressional Districts for Baltimore City}
\format{
A data frame with 3 rows and 15 variables:
\describe{
  \item{statefp}{...}
  \item{cd116fp}{...}
  \item{geoid}{GeoID}
  \item{namelsad}{...}
  \item{lsad}{...}
  \item{cdsessn}{...}
  \item{mtfcc}{...}
  \item{funcstat}{...}
  \item{aland}{Land area}
  \item{awater}{Water area}
  \item{intptlat}{Interior center point latitude}
  \item{intptlon}{Interior center point longitude}
  \item{label}{Congressional District label}
  \item{name}{Congressional District name}
  \item{geometry}{Multipolygon with Congressional district boundary}
  ...
}
}
\source{
\url{...}
}
\usage{
congressional_districts
}
\description{
U.S. Congressional Districts overlapping with Baltimore City. Downloaded with the tigris package.
}
\keyword{datasets}
