% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{council_districts}
\alias{council_districts}
\title{Baltimore City Council Districts}
\format{
A data frame with 14 rows and 2 variables:
\describe{
  \item{id}{Number of the City Council district}
  \item{name}{Name of the City Council district}
  \item{geometry}{Multipolygons for boundaries of City Council districts}
  ...
}
}
\source{
\url{https://geodata.baltimorecity.gov/egis/rest/services/CityView/City_Council_Districts/MapServer/0}
}
\usage{
council_districts
}
\description{
...
}
\keyword{datasets}
