% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csas}
\alias{csas}
\title{Baltimore City Community Statistical Areas}
\format{
A data frame with 55 rows and 3 variables:
\describe{
  \item{\code{id}}{Community Statistical Area identification number}
  \item{\code{name}}{Community Statistical Area name}
  \item{\code{geometry}}{Multipolygon with area boundary}
}
}
\source{
\url{...}
}
\usage{
csas
}
\description{
Baltimore City Community Statistical Areas (CSAs)
}
\keyword{datasets}
