% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase.R
\name{erase_baltimore}
\alias{erase_baltimore}
\alias{erase_baltimore_park}
\alias{erase_baltimore_parks}
\alias{erase_baltimore_water}
\title{Erase any are overlapping the geometry}
\usage{
erase_baltimore_parks(x, ...)

erase_baltimore_water(x, water = c("city", "msa"), ...)
}
\arguments{
\item{x}{A \code{sf}, \code{sfc}, or \code{bbox} object to erase or trim.}

\item{...}{
  Arguments passed on to \code{\link[sfext:st_erase]{sfext::st_erase}}
  \describe{
    \item{\code{flip}}{If \code{TRUE}, use \code{\link[sf:geos_binary_ops]{sf::st_intersection()}} to "erase" geometry of x
that intersects y; if \code{FALSE} use \code{\link[sf:geos_binary_ops]{sf::st_difference()}} to trim x to y
geometry, Default: \code{FALSE}.}
    \item{\code{union}}{If \code{TRUE}, union \code{y} with \code{\link[sf:geos_combine]{sf::st_union()}} before applying
difference/intersection; defaults to \code{TRUE}.}
    \item{\code{combine}}{If \code{TRUE} and \code{union = TRUE}, combine \code{y} with
\code{\link[sf:geos_combine]{sf::st_combine()}} before unioning. Defaults to \code{FALSE}.}
  }}

\item{water}{Type of water data to use, either \code{baltimore_water} or
\code{baltimore_msa_water}.}
}
\description{
Erase any are overlapping the geometry
}
