% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{expand_limits_to_area}
\alias{expand_limits_to_area}
\title{Expand limits of ggplot map to a selected area}
\usage{
expand_limits_to_area(area, crs = 2804)
}
\arguments{
\item{area}{sf object.}

\item{crs}{EPSG code for the coordinate reference system for the plot. Default is 2804. See \url{https://epsg.io/}}
}
\description{
Gets the bounding box of an area and passes the coordinates to the \code{ggplot2::coord_sf} function. This function is useful for highlighting a defined area within a plot or expanding a plot to make space for labels and/or annotation.
}
