% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{explore_baltimore}
\alias{explore_baltimore}
\title{Stories from Explore Baltimore Heritage}
\format{
A data frame with 459 rows and 10 variables:
\describe{
  \item{id}{Story identifier}
  \item{featured}{Featured indicator}
  \item{modified}{Modified date/time}
  \item{title}{Story title}
  \item{address}{Street address for story location}
  \item{thumbnail}{URL for thumbnail-size featured image}
  \item{fullsize}{URL for full-size featured image}
  \item{url}{URL for story}
  \item{geometry}{Point for story location}
  ...
}
}
\source{
\url{https://explore.baltimoreheritage.org/}
}
\usage{
explore_baltimore
}
\description{
A table of public stories on the Explore Baltimore Heritage website.
}
\keyword{datasets}
