% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maryland_open_data.R
\name{get_area_crashes}
\alias{get_area_crashes}
\title{Get vehicle crashes for area in Baltimore from Maryland Open Data portal}
\usage{
get_area_crashes(
  area,
  start_year = 2020,
  end_year = NULL,
  geometry = FALSE,
  trim = FALSE
)
}
\arguments{
\item{area}{\code{\link{sf}} object.}

\item{start_year}{earliest year of crash data to return.}

\item{end_year}{latest year of crash data to return. If \code{end_year} is not provided, only a single year is returned.}

\item{geometry}{If TRUE, return a sf object. Default FALSE.}

\item{trim}{If TRUE, data trimmed to area with \code{\link{sf::st_intersection()}}. Default FALSE.}
}
\description{
Get vehicle crashes for selected area in Baltimore City.
}
