% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_crime.R
\name{get_area_crime}
\alias{get_area_crime}
\title{Get area crimes from Open Baltimore}
\usage{
get_area_crime(
  area,
  description = NULL,
  date_range = NULL,
  where = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  unit = "m",
  trim = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804)
)
}
\arguments{
\item{area}{sf, sfc, or bbox object. If multiple areas are provided, they are
unioned into a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}.}

\item{description}{Crime type or description. Supported options include "AGG.
ASSAULT", "ARSON", "AUTO THEFT", "BURGLARY", "COMMON ASSAULT", "HOMICIDE",
"LARCENY", "LARCENY FROM AUTO", "RAPE", "ROBBERY - CARJACKING", "ROBBERY -
COMMERCIAL", "ROBBERY - RESIDENCE", "ROBBERY - STREET", or "SHOOTING". Not
case sensitive.}

\item{date_range}{Date range as character vector in format of c("YYYY-MM-DD",
"YYYY-MM-DD"). Minimum and maximum values are used if length is greater
than 1.}

\item{where}{where query string passed to esri2sf, Default: \code{NULL}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{trim}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}

\item{crs}{Cordinate reference system to return, Default: 4326 for
\code{\link[sfext:sf_to_df]{sf_to_df()}} and \code{NULL} for \code{\link[sfext:df_to_sf]{df_to_sf()}}.}
}
\description{
Get reported crimes since 2014 for a specific area.
}
\examples{
\dontrun{
# Get shootings for the Lauraville area
area <- get_area("neighborhood", "Barclay")
crimes <- get_area_crime(
  area = area,
  date_range = c("2022-01-01", "2022-12-31"),
  description = "SHOOTING"
)
}
}
