% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area_data}
\alias{get_area_data}
\title{Get local or cached data for an area}
\usage{
get_area_data(
  data,
  area,
  extdata = NULL,
  diag_ratio = NULL,
  dist = NULL,
  crop = TRUE,
  trim = FALSE,
  crs = NULL
)
}
\arguments{
\item{data}{sf object including data in area}

\item{area}{sf object. If multiple areas are provided, they are unioned into a single sf object using \code{\link[sf]{st_union()}}}

\item{extdata}{Character. Name of cached geopackage file to load data from.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{crop}{If TRUE, data cropped to area (or buffered area if dist or diag_ratio are provided) \code{\link[sf]{st_crop()}}. Default TRUE.}

\item{trim}{If TRUE, data trimmed to area (or buffered area if dist or diag_ratio are provided) with \code{\link[sf]{st_intersection()}}. Default FALSE.}

\item{crs}{Selected CRS for returned data}
}
\description{
Returns data for a selected area or areas with an optional buffer.
If both crop and trim are FALSE, the function uses \code{\link[sf]{st_join()}} to return provided data without any changes to geometry.
}
