% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area_esri_data}
\alias{get_area_esri_data}
\title{Get GIS data from an ArcGIS FeatureServer or MapServer}
\usage{
get_area_esri_data(
  area = NULL,
  bbox = NULL,
  type = c("md food stores 2017 2018", "farmers markets 2020",
    "baltimore food stores 2016", "baltimore demolitions", "contour 2ft",
    "contours 10ft", "open vacant building notices", "liquor licenses",
    "fixed speed cameras", "red light cameras", "edge of pavement"),
  trim = FALSE,
  crs = 2804
)
}
\arguments{
\item{area}{sf object. Optional. Only used if trim is TRUE.}

\item{bbox}{bbox object. Optional but suggested to avoid downloading entire layer. See sf::st_bbox() for more information.}

\item{type}{Type of data to get. Options include "md food stores 2017 2018", "farmers markets 2020", "baltimore food stores 2016", "baltimore demolitions", "contour 2ft", "contours 10ft", "open vacant building notices", "liquor licenses", "fixed speed cameras", "red light cameras", and "edge of pavement"}

\item{trim}{Logical. Default FALSE. If TRUE, area is required.}
}
\description{
Wraps the esri2sf::esri2sf() function to download an ArcGIS FeatureServer or MapServer.
Some of the data (e.g. Liquor Licenses) is missing data important data.
}
