% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_area_mask}
\alias{get_area_mask}
\title{Get a mask for an area}
\usage{
get_area_mask(area, edge = NULL, crs = 2804)
}
\arguments{
\item{area}{sf object. If multiple areas are provided, they are unioned into a single sf object using \code{\link[sf]{st_union()}}}

\item{edge}{sf object. Must match CRS of area. Defaults to bounding box of buffered area, converted to an sf object.}

\item{crs}{Selected CRS for returned mask.}
}
\description{
Returns a mask for an area or areas as an sf object. Used by the \code{map_area_with_snapbox} function.
}
