% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_streetnames.R
\name{get_area_streets}
\alias{get_area_streets}
\title{Get area streets}
\usage{
get_area_streets(area, street_type = NULL, sha_class = NULL)
}
\arguments{
\item{area}{sf object. Returned streets are cropped to area}

\item{street_type}{selected street subtypes to include. Includes all subtypes except alleys ("STRALY") by default.
Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN")}

\item{sha_class}{selected SHA classifications to include.
"all" selects all streets with an assigned SHA classification (around one-quarter of all street segments).
Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")}
}
\description{
Get streets within an area or areas.
}
