% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning.R
\name{get_area_zoning}
\alias{get_area_zoning}
\title{Get zoning data a local area}
\usage{
get_area_zoning(
  area,
  category = c("all", "residential", "commercial", "industrial"),
  diag_ratio = 0.125,
  dist = NULL,
  crop = TRUE,
  trim = FALSE,
  crs = NULL
)
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{category}{Zoning category to return. "all", "residential", "commercial", "industrial"}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{crop}{If TRUE, data cropped to area or bounding box \code{\link[sf]{st_crop()}} adjusted by the `dist`, `diag_ratio`, and `asp` provided. Default `TRUE`.}

\item{trim}{If TRUE, data trimmed to area with \code{\link[sf]{st_intersection()}}. This option is not supported for any adjusted areas that use the `dist`, `diag_ratio`, or `asp` parameters. Default `FALSE`.}

\item{crs}{Coordinate Reference System (CRS) to use for the returned data. The CRS of the provided data and bounding box or area must match one another but are not required to match the CRS provided by this parameter.}
}
\description{
Get zoning codes for an area within the city.
The 2017 zoning data does not include any exemptions granted by the BMZA (Board of Municipal Zoning Appeals).
}
