% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_asp_adj_bbox}
\alias{get_asp_adj_bbox}
\title{Get area bounding box adjusted to aspect ratio}
\usage{
get_asp_adj_bbox(area = NULL, bbox = NULL, asp = NULL)
}
\arguments{
\item{area}{sf object}

\item{bbox}{bbox object}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").}
}
\value{
Class bbox object
}
\description{
Get area bounding box adjusted to aspect ratio
}
\details{
Takes an area or bounding box and returns a bounding box that matches
the provided aspect ratio and contains the area or bounding box provided.
Common aspect ratios include "1:1" (1), "4:6" (0.666), "8.5:11", "16:9" (1.777).
The asp parameter supports both numeric values and ratios matching the
format of "width:height".
}
