% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_baltimore_data.R
\name{get_citations}
\alias{get_citations}
\title{Get selected Environmental Control Board citations from Open Baltimore data portal}
\usage{
get_citations(
  violation_type = NULL,
  start_date = NULL,
  end_date = NULL,
  filter_by = c("neighborhood", "council", "police"),
  area = NULL,
  geometry = FALSE
)
}
\arguments{
\item{start_date}{The start date of the time period during which downloaded violations were cited in the format "YYYY-MM-DD". An end date must be provided if a start date is provided.}

\item{end_date}{The end date of the time period during which downloaded violations were cited in the format "YYYY-MM-DD".  An start date must be provided if an end date is provided.}

\item{filter_by}{A character vector with the type of area to filter citations by. Supported area types include neighborhoods ("neighborhood"), City Council districts ("council"), and Police Districts ("police").}

\item{area}{A character vector with a neighborhood name, Council district number, or Police district name. Filtering by multiple areas or multiple area types is not currently supported.}

\item{geometry}{If TRUE the service requests are converted to an sf object with a projected CRS (2804). Requests with no coordinates are excluded if TRUE. Defaults to FALSE.}

\item{request_type}{Character vector matching one of the possible \code{violation_types}.}
}
\description{
This function uses the RSocrata package to download citations from the Open Baltimore data portal.
Requests may be one or more service request types, by date, or by neighborhood, City Council District, or Police District.
Filtering by multiple criteria or limiting the period of time covered is recommended to avoid a long wait for a large data file to download.
}
\examples{

\dontrun{
## Get all citations for the Northern Police District for 2018 and 2019
get_citations(
  filter_by = "police",
  area = "Northern",
  start_date = "2018-01-01",
  end_date = "2019-12-31"
)
}

\dontrun{
## Get housing citations in Hampden for high grass/weeds between April and October 2019
get_citations(
  violation_type = "HIGH GRASS AND WEEDS",
  filter_by = "neighborhood",
  area = "Hampden",
  start_date = "2019-04-01",
  end_date = "2019-10-31",
  geometry = TRUE
) ## Return an sf object
}
}
