% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearby_areas.R
\name{get_nearby_areas}
\alias{get_nearby_areas}
\title{Get nearby areas}
\usage{
get_nearby_areas(
  area,
  type = c("neighborhood", "council district", "legislative district",
    "congressional district", "planning district", "police district", "csa",
    "park district"),
  dist = 1,
  exclude_area = TRUE,
  residential = FALSE
)
}
\arguments{
\item{area}{sf object. Must have a name column for exclude_area to work.}

\item{type}{Required. Supported values include "neighborhood", "council
district", "legislative district", "congressional district", "planning
district", "police district", "csa", and "park district". The type may be
different than the type of the area provided.}

\item{dist}{Distance in meters for matching nearby areas. Default is 1 meter.}

\item{exclude_area}{Logical. Default TRUE. If FALSE, include the same areas
provided to area (assuming the areas provide are the same type as the
parameter provided to get_nearby_areas).}

\item{residential}{Logical. Default FALSE. If the type is neighborhood, set
TRUE to only return residential neighborhoods (excluding industrial areas,
business parks, and parks/reservoirs).}
}
\description{
Return areas of a selected type within a set distance of another area.
}
