% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_baltimore_data.R
\name{get_open_baltimore_resource}
\alias{get_open_baltimore_resource}
\title{Get dataset from Open Baltimore with optional SoQL parameters}
\usage{
get_open_baltimore_resource(
  resource = NULL,
  select = NULL,
  where = NULL,
  query = NULL,
  geometry = FALSE
)
}
\arguments{
\item{resource}{Character vector with the Socrata dataset identifier for the dataset on Open Baltimore.}

\item{select}{Character vector with SODA $select parameter. This is the set of columns to be returned, similar to a SELECT in SQL. \url{https://dev.socrata.com/docs/queries/select.html}}

\item{where}{Character vector with SODA $where parameter. Filters the rows to be returned, similar to WHERE. \url{https://dev.socrata.com/docs/queries/where.html}}

\item{query}{Character vector with SODA $query parameter. A full SoQL query string, all as one parameter. \url{https://dev.socrata.com/docs/queries/query.html}}

\item{geometry}{If this is set to TRUE and dataset contains a 'longitude' and 'latitude' column, the function returns a \code{sf} object.}
}
\description{
Get a selected Open Baltimore dataset filtered by selected Socrata Query Language (SoQL) parameters as a tibble or sf object. Details on SoQL queries are found in the Socrata API documentation \url{https://dev.socrata.com/docs/queries/}
}
\examples{

\dontrun{
## Get  (outdated) list of city department names, department head titles, and department head names
get_open_baltimore_resource(
  resource = "cut3-c4bx",
  select = "formal_agency_name, agency_head, agency_head_title, agency_type",
  where = "agency_type like 'Department'"
)
}
}
