% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_osm_feature}
\alias{get_osm_feature}
\title{Get OSM feature}
\usage{
get_osm_feature(
  area,
  key,
  value,
  osm_return = c("osm_points", "osm_lines", "osm_polygons", "osm_multilines",
    "osm_multipolygons"),
  trim = FALSE,
  crs = 4326
)
}
\arguments{
\item{area}{sf object. If multiple areas are provided, they are unioned into a single sf object using \code{sf::st_union()}}

\item{key}{feature key for overpass query}

\item{value}{for feature key; can be negated with an initial exclamation mark, value = "!this", and can also be a vector, value = c ("this", "that").}

\item{osm_return}{Character vector length 1 with geometry type to return. Defaults to returning all types.}

\item{trim}{Logical. Default FALSE. If TRUE, use the \code{sf::st_intersection()} function to trim results to area polygon instead of bounding box.}

\item{crs}{EPSG code for the coordinate reference system for the plot. Default is 2804. See \url{https://epsg.io/}}
}
\description{
Wraps \code{osmdata} functions.
}
