% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_baltimore_data.R
\name{get_permits}
\alias{get_permits}
\title{Get selected housing permits from Open Baltimore data portal}
\usage{
get_permits(
  area,
  permit_type = NULL,
  start_date = NULL,
  end_date = NULL,
  filter_by = c("neighborhood", "police_district", "council_district"),
  area_name = NULL,
  geometry = FALSE
)
}
\arguments{
\item{area}{An sf object with a 'name' column. If area is provided, the function returns data within the bounding box of the provided area or areas.}

\item{start_date}{The start date of the time period during which permits were issued in the format "YYYY-MM-DD". An end date must be provided if a start date is provided.}

\item{end_date}{The end date of the time period during which permits were issued in the format "YYYY-MM-DD".  An start date must be provided if an end date is provided.}

\item{filter_by}{A character vector with the type of area to filter permits by. Supported area types include neighborhoods ("neighborhood"), City Council districts ("council_district"), and Police Districts ("police_district").}

\item{area_name}{A character vector with a neighborhood name or Police district name. Filtering by multiple areas or multiple area types is not currently supported.}

\item{geometry}{If TRUE the returned permit data is converted to an sf object with a projected CRS (2804). Permits with no coordinates are excluded if TRUE. Defaults to FALSE.}

\item{request_type}{Character vector matching one of the possible permits types ("USE", "DEM", "COM").}
}
\description{
This function uses the RSocrata package to download Housing Permits from the Open Baltimore data portal.
Requests may be filtered by permit type, by date, or by neighborhood, Council District, or Police District.
}
\examples{

\dontrun{
## Get all demolition permits for Council District 9 between 2015 and 2019
get_permits(
  permit_type = "DEM",
  start_date = "2015-01-01",
  end_date = "2019-12-31",
  filter_by = "council_district",
  area_name = "9"
)
}

\dontrun{
## Get all housing permits for the Mount Washington neighborhood from Dec. 2019 to Jan. 2020
get_permits(
  start_date = "2019-12-01",
  end_date = "2020-01-31",
  filter_by = "neighborhood",
  area_name = "MOUNT WASHINGTON"
)
}
}
