% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_property.R
\name{get_real_property}
\alias{get_real_property}
\title{Get real property data for an area}
\usage{
get_real_property(
  area,
  area_type = c("neighborhood", "council_district", "police_district", "csa",
    "block_group", "tract"),
  buffer = FALSE,
  buffer_distance = 0
)
}
\arguments{
\item{area}{Required sf class tibble. Must include a name column.}

\item{area_type}{Optional character vector for the type of area (e.g. "neighborhood", "block_group"). The name column for the provided \code{area} must match the names or geoid of the one or more areas of the provided type to return data.
If buffer is TRUE, the area_type is ignored and the \code{sf::st_crop} function is used.}

\item{buffer}{Logical. If TRUE, the returned real property data includes property within a default buffered distance (1/8th of the diagonal distance across the bounding box) or within the provided \code{buffer_distance}.
If FALSE, the returned real property data includes property cropped to the bounding box of the area.}

\item{buffer_distance}{A single numeric vector with the buffer distance in meters.}
}
\description{
Map showing parcels described as owner occupied, non-owner occupied, vacant, and unimproved.
Real property or parcel data is from the Maryland State Department of Assessment and Taxation and may include errors. Check the \code{mapbaltimore::real_property} data description for details.
}
