% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streets.R
\name{get_streets}
\alias{get_streets}
\title{Get streets}
\usage{
get_streets(
  street_name = NULL,
  exclude_name = NULL,
  street_type = NULL,
  sha_class = NULL,
  block_num = NULL,
  bbox = NULL,
  union = TRUE
)
}
\arguments{
\item{street_name}{Street names to return. Required.}

\item{exclude_name}{Street names to exclude}

\item{street_type}{selected street subtypes to include. By default, the returned data includes all subtypes except alleys ("STRALY"). Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN"). Not supported for}

\item{sha_class}{selected SHA classifications to include. "all" selects all streets with an assigned SHA classification (around one-quarter of all street segments). Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")}

\item{block_num}{Integer vector with block number c(300) or range of block numbers (c(100, 500)) to filter streets.}

\item{bbox}{bbox to crop returned streets. Optional.}

\item{union}{Logical. If `TRUE`, use \code{st_union} to combine geometry by
`fullname` of the streets.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get streets in Baltimore City by name with option to exclude streets by name,
crop to a bounding box, or to filter to selected street types or functional
classifications.
}
\details{
DETAILS
}
\examples{
\dontrun{
get_streets(street_name = "UNIVERSITY PKWY")

get_streets(street_name = c("E FAYETTE", "ORLEANS"), block_num = c(1700, 3600))
}
}
\seealso{
get_area_streets

\code{\link[mapbaltimore]{streets}}
 \code{\link[mapbaltimore]{get_area_streets}}
}
