% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoning.R
\name{get_zoning}
\alias{get_zoning}
\title{Get zoning data a local area}
\usage{
get_zoning(area, dist = NULL)
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{dist}{If default (NULL), the returned real property data includes property within a default buffered distance (1/8th of the diagonal distance across the bounding box). If numeric, the function returns data cropped to area buffered by this distance in meters.}
}
\description{
Get zoning codes for an area within the city.
The 2017 zoning data does not include any exemptions granted by the BMZA (Board of Municipal Zoning Appeals).
}
