% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inspire_plans}
\alias{inspire_plans}
\title{INSPIRE Plans}
\format{
A data frame with 24 rows and 23 variables:
\describe{
\item{\code{plan_name}}{Plan name}
\item{\code{plan_name_short}}{Plan name (short)}
\item{\code{overall_status}}{Overall status}
\item{\code{inspire_lead_planner}}{Lead INSPIRE Planner}
\item{\code{plan_url}}{Baltimore City Department of Planning plan webpage url}
\item{\code{year_launched}}{Year launched}
\item{\code{year_adopted}}{Year adopted by Planning Commission}
\item{\code{adoption_status}}{Planning Commission adoption status}
\item{\code{adoption_date}}{Planning Commission adoption data}
\item{\code{document_url}}{Adopted plan PDF url}
\item{\code{recommendation_report_status}}{Recommendation report status}
\item{\code{recommendation_report_url}}{Draft recommendation report PDF url}
\item{\code{kick_off_presentation_date}}{Kick-off presentation date}
\item{\code{launch_date_target}}{Target launch date}
\item{\code{walking_route_id_target_date}}{Primary walking route identification date}
\item{\code{recommendations_date_target}}{Target draft recommendation report publication date}
\item{\code{commission_review_date_target}}{Target Planning Commission review date}
\item{\code{implementation_status}}{Plan implementation status}
\item{\code{program_numbers}}{School program numbers}
\item{\code{planning_districts}}{Planning Districts}
\item{\code{neighborhoods}}{Neighborhoods}
\item{\code{council_districts}}{Baltimore City Council Districts}
\item{\code{geometry}}{MULTIPOLYGON boundary geometry}
}
}
\usage{
inspire_plans
}
\description{
Data frame and boundary geometry for INSPIRE Plans adopted and in progress.
Learn more about the INSPIRE program: \url{https://planning.baltimorecity.gov/planning-inspire}
}
\details{
Last updated: 2025-09-03
}
\keyword{datasets}
