% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streets.R
\name{label_area_streets}
\alias{label_area_streets}
\title{Label area street names at selected locations}
\usage{
label_area_streets(
  area,
  geom = c("label", "label_repel"),
  sha_class = NULL,
  label_location = c("area", "edge", "topright", "bottomleft")
)
}
\arguments{
\item{area}{sf object. Labels returned for streets in or around area bounding box.}

\item{geom}{Character vector matching name of geom returned with labels. "label" for "ggplot::geom_label" or "label_repel" for "ggrepel::geom_label_repel" are supported.}

\item{sha_class}{Character vector. "all" selects all streets with an assigned SHA classification. Options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")}

\item{label_location}{Options include "area", "edge", "topright", or "bottomleft". Defaults to "area"}
}
\description{
Label street names at selected locations.
}
