% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_area_data.R
\name{layer_area_data}
\alias{layer_area_data}
\title{Layer for area data}
\usage{
layer_area_data(
  area = NULL,
  bbox = NULL,
  data = NULL,
  extdata = NULL,
  cachedata = NULL,
  asis = FALSE,
  diag_ratio = NULL,
  dist = NULL,
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  crs = 2804,
  mapping = aes(),
  inherit.aes = TRUE,
  show_area = FALSE,
  area_color = "gray30",
  area_fill = NA,
  mask = FALSE,
  ...
)
}
\arguments{
\item{area}{`sf` object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf]{st_union}}}

\item{bbox}{`bbox` object defining area used to filter data. If an area is
provided, the bounding box is ignored.}

\item{data}{`sf` object including data in area}

\item{extdata}{Character. Name of an external geopackage (.gpkg) file
included with the package where selected data is available. Available data
includes "trees", "unimproved_property", and "vegetated_area"}

\item{cachedata}{Character. Name of a cached geopackage (.gpkg) file where
selected data is available. Running \code{cache_mapbaltimore_data()} caches data
for "real_property", "baltimore_msa_streets", and "edge_of_pavement"}

\item{asis}{Logical. Default FALSE. If TRUE, use inherited data as is without cropping to area.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crop}{If TRUE, data cropped to area or bounding box
\code{\link[sf]{st_crop}} adjusted by the `dist`, `diag_ratio`, and `asp`
provided. Default `TRUE`.}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf]{st_intersection}}. This option is not supported for any
adjusted areas that use the `dist`, `diag_ratio`, or `asp` parameters.
Default `FALSE`.}

\item{crs}{Coordinate Reference System (CRS) to use for the returned data.
The CRS of the provided data and bounding box or area must match one
another but are not required to match the CRS provided by this parameter.}

\item{show_area}{Logical. Default FALSE. If TRUE, add an outline of the area
to the layer.}

\item{area_color}{Character to set fixed color aesthetic for area layer.}

\item{area_fill}{to set fixed fill aesthetic for area layer.}

\item{mask}{Logical. Default FALSE. If TRUE, add a mask using
\code{layer_area_mask}}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsf]{ggplot2::geom_sf}}, \code{\link[ggplot2:ggsf]{ggplot2::geom_sf}}
  \describe{
    \item{\code{mapping}}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}
    \item{\code{inherit.aes}}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
  }}
}
\description{
Combines get_area_data and geom_sf into a single call. Optionally can outline
area and call layer_area_mask.
}
