% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_area_streets.R
\name{layer_area_streets}
\alias{layer_area_streets}
\title{Add a layer to a gpplot2 map with area streets, street names, or both}
\usage{
layer_area_streets(
  area = NULL,
  street_type = NULL,
  sha_class = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  msa = FALSE,
  hide = c("names", "streets", "none"),
  name_location = NULL,
  ...
)
}
\arguments{
\item{area}{sf object. Return}

\item{street_type}{selected street subtypes to include. By default, the returned data includes all subtypes except alleys ("STRALY").
Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN")}

\item{sha_class}{selected SHA classifications to include.
"all" selects all streets with an assigned SHA classification (around one-quarter of all street segments).
Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{Logical. Default `FALSE`. Trim streets to area using `sf::st_intersection()`.}

\item{msa}{Logical. Default `FALSE`. Get streets from cached `baltimore_msa_streets.gpkg` file using `cachedata` parameter of `get_area_data` function.}

\item{hide}{Options include c("names", "streets", "none"). Defaults to "names"}

\item{name_location}{Options include c("area", "edge", "topright", or "bottomleft"). Defaults to NULL.}

\item{...}{Other parameters to pass along to `ggplot2::geom_sf()` that maps the streets.}
}
\description{
Add a layer to a gpplot2 map with area streets, street names, or both.
}
