% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_streets}
\alias{main_streets}
\title{Baltimore City Main Streets}
\format{
A data frame with 10 rows and 7 variables:
\describe{
\item{\code{id}}{Main Street ID from source FeatureLayer}
\item{\code{name}}{Main Street name}
\item{\code{name_abb}}{Name abbreviation}
\item{\code{url}}{Main Street partner organization URL}
\item{\code{funding_status}}{Funding status (active or inactive)}
\item{\code{name_short}}{Short name}
\item{\code{geometry}}{sfc list column with MULTIPOLYGON boundary geometry}
}
}
\source{
\url{https://services1.arcgis.com/43Lm3JYE3nM91DAF/arcgis/rest/services/MainStreets/FeatureServer/0}
}
\usage{
main_streets
}
\description{
Boundaries for Baltimore City Main Street programs, including two programs
that are not currently funded but formerly participated in the program.
}
\keyword{datasets}
