% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_bcps_programs.R
\name{map_area_bcps_programs}
\alias{map_area_bcps_programs}
\title{Map BCPS programs and attendance zones for a local area}
\usage{
map_area_bcps_programs(area)
}
\arguments{
\item{area}{sf object}
}
\description{
Map showing BCPS school zones that overlap with a provided area or areas.
If the area sf tibble includes multiple areas, a separate map is created for each area provided.
}
\examples{

\dontrun{
## Map school attendance boundary zones for the Madison Park neighborhood
madisonpark <- get_area(area_type = "neighborhood", area_name = "Madison Park")
map_area_bcps_programs(area = madisonpark)
}

\dontrun{
## Map school attendance boundary zones for City Council District 2
district9 <- get_area(
  type = "council district",
  area_name = "9"
)
map_area_bcps_programs(area = district9)
}
}
