% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcps_zones.R
\name{map_area_bcps_zones}
\alias{map_area_bcps_zones}
\title{Map BCPS School Attendance zones for a local area}
\usage{
map_area_bcps_zones(area, crop = "none")
}
\arguments{
\item{area}{Name of the neighborhood in sentence case.}

\item{crop}{Default is "none." Only supported option is "standard" which uses the get_buffered_area function to crop the BCPS zones.}
}
\description{
Map showing BCPS school zones that overlap with a provided area or areas. If the area sf tibble includes multiple areas, a separate map is created for each area provided.
}
\examples{

\dontrun{
## Map school attendance boundary zones for the Madison Park neighborhood
madisonpark <- get_area(area_type = "neighborhood", area_name = "Madison Park")
map_area_bcps_zones(area = madisonpark)
}

\dontrun{
## Map school attendance boundary zones for City Council District 2
district9 <- get_area(area_type = "council_district", area_name = "9")
map_area_bcps_zones(area = district9)
}
}
