% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_highlighted.R
\name{map_area_highlighted}
\alias{map_area_highlighted}
\title{Maps a highlighted area within the context of multiple areas}
\usage{
map_area_highlighted(area, highlight_name = "all")
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{highlight_name}{Character vector. Required. Use "all" to create a grid of maps highlighting each area in the provided sf object or provide the name of one or more areas to highlight.}
}
\description{
Map highlighting the location of an area the context of multiple areas.
}
