% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_in_city.R
\name{map_area_in_city}
\alias{map_area_in_city}
\title{Maps an area in the context of the city}
\usage{
map_area_in_city(area, area_label = NULL, map_title = NULL)
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{area_label}{Optional character vector of any length. Replaces the name column in the provided area sf object as the label and in the title (if map_title is not provided).}

\item{map_title}{Optional character vector of length 1 to replace the default map title.}
}
\description{
Map showing the location of an area within the city.
}
\examples{

\dontrun{
## Area with a defined label
district2 <- get_area(
type =  "council district",
name =  "2")

map_area_in_city(
area = district2,
area_label = "Council District 2")
}

\dontrun{
## Multiple areas in a single map
selected_se_neighborhoods <- get_area(
type =  "neighborhood",
name =  c("Upper Fells Point", "Fells Point", "Canton"))

map_area_in_city(area = selected_se_neighborhoods)
}

\dontrun{
## Area with a defined map title
canton_industrial <- get_area(
type =  "neighborhood",
name =  "Canton Industrial Area")

map_area_in_city(
area = canton_industrial,
map_title = "Canton Industrial Area is the largest neighborhood areas in Baltimore")
}

}
