% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_mta_services.R
\name{map_area_mta_services}
\alias{map_area_mta_services}
\title{Map MTA services}
\usage{
map_area_mta_services(
  area,
  mta_services = "bus_lines",
  diag_ratio = 0.166,
  asp = NULL
)
}
\arguments{
\item{area}{sf object. Required.}

\item{mta_services}{Character vector. Default is "bus_lines" to use mta_bus_lines data.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}
}
\description{
Map MTA services. MTA bus lines are currently the only supported service.
}
