% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_property.R
\name{map_area_property}
\alias{map_area_property}
\title{Map real property data for an area}
\usage{
map_area_property(
  area,
  property = c("improved", "vacant", "principal residence", "use", "building type",
    "value"),
  dist = NULL,
  diag_ratio = 0.1,
  asp = NULL,
  trim = FALSE,
  mask = FALSE
)
}
\arguments{
\item{area}{Simple features object. Function currently supports only a single area at a time.}

\item{property}{Real property variable to map. Options include c("improved", "vacant", "principal residence", "value"). Currently supports only one variable at a time.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").}

\item{trim}{If TRUE, data trimmed to area with \code{\link[sf]{st_intersection()}}. This option is not supported for any adjusted areas that use the `dist`, `diag_ratio`, or `asp` parameters. Default `FALSE`.}

\item{mask}{If `TRUE`, apply a white, 0.6 alpha mask over property located outside the provided area. Default `FALSE.`}
}
\description{
Map real property data by improvement, vacancy, principal residence status, and other characteristics.
This function is intended to replace `map_tenure`, `map_vacancy`, and (eventually) `map_decade_built.`
Real property or parcel data is from the Maryland State Department of Assessment and Taxation and may include outdated or inaccurate information.
}
