% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_streetnames.R
\name{map_area_streetnames}
\alias{map_area_streetnames}
\title{Map area with street names labelled at selected locations}
\usage{
map_area_streetnames(
  area,
  label_location = c("area", "edge", "topright", "bottomleft"),
  sha_class = TRUE
)
}
\arguments{
\item{area}{sf class tibble. Object must include a name column.}

\item{label_location}{Character vector. Must be "area", "edge", "topright", or "bottomleft"}

\item{sha_class}{Logical. Option to label all streets (FALSE) or only State Highway Administration classified streets (TRUE). Default is TRUE.}
}
\description{
Map streets within an area or areas with street name labels at selected locations.
}
