% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_in_city.R
\name{map_area_with_snapbox}
\alias{map_area_with_snapbox}
\title{Maps an area or areas using the snapbox package}
\usage{
map_area_with_snapbox(area, map_style = stylebox::mapbox_satellite_streets())
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{map_style}{Required. \code{\link{stylebox}} function referencing mapbox map styles. Default is \code{\link[stylebox]{mapbox_satellite_streets()}}}
}
\description{
Map an area or areas using the \code{\link{snapbox}} package.
}
