% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning.R
\name{map_area_zoning}
\alias{map_area_zoning}
\title{Map zoning for a local area}
\usage{
map_area_zoning(
  area,
  category = c("all", "residential", "commercial", "industrial"),
  diag_ratio = 0.125
)
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{category}{"all", "residential", "commercial", or "industrial"}

\item{diag_ratio}{}
}
\description{
Map zoning/zoning overlay codes for an area within the city.
The 2017 zoning data does not include any exemptions granted by the BMZA (Board of Municipal Zoning Appeals).
}
\examples{
\dontrun{
## Map zoning code for Bayview neighborhood
bayview <- get_area(type = "neighborhood", area_name = "Bayview")
map_zoning(area = bayview)
}
}
