% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_bcps_zones.R
\name{map_bcps_zones}
\alias{map_bcps_zones}
\title{Map BCPS School Attendance zones for a local area}
\usage{
map_bcps_zones(neighborhood_label, neighborhood_color = "gray20")
}
\arguments{
\item{neighborhood_label}{Name of the neighborhood in sentence case.}

\item{neighborhood_color}{Color of the neighborhood boundary line}
}
\description{
Map showing the school zones that overlap with the selected neighborhood.
}
