% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_decade_built.R, R/real_property.R
\name{map_decade_built}
\alias{map_decade_built}
\title{Map real property data to show decade built}
\usage{
map_decade_built(area = NULL)

map_decade_built(area = NULL)
}
\arguments{
\item{area}{sf class tibble. Object must include a name column.}

\item{neighborhood_label}{Required. Name of the neighborhood in sentence case}

\item{neighborhood_color}{Optional. Color of the neighborhood boundary line}
}
\description{
Map showing parcels described by decade built.
Parcel data is from the Maryland State Department of Assessment and Taxation and may include errors.

Map showing parcels described as owner occupied, non-owner occupied, vacant, and unimproved.
Parcel data is from the Maryland State Department of Assessment and Taxation and may include errors.
}
