% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_neighborhood_in_city.R
\name{map_neighborhood_in_city}
\alias{map_neighborhood_in_city}
\title{Maps a neighborhood in the context of the city}
\usage{
map_neighborhood_in_city(neighborhood_label, neighborhood_color = "gray20")
}
\arguments{
\item{neighborhood_label}{...}

\item{neighborhood_color}{...}
}
\description{
Map showing the location of the neighborhood within the city.
}
