% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tenure.R
\name{map_tenure}
\alias{map_tenure}
\title{Map real property data to show tenure for neighborhood}
\usage{
map_tenure(neighborhood_label, neighborhood_color = "gray20")
}
\arguments{
\item{neighborhood_label}{Required. Name of the neighborhood in sentence case}

\item{neighborhood_color}{Optional. Color of the neighborhood boundary line}
}
\description{
Map showing parcels described as owner occupied, non-owner occupied, vacant, and unimproved.
Parcel data is from the Maryland State Department of Assessment and Taxation and may include errors.
}
