% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_property.R
\name{map_vacancy}
\alias{map_vacancy}
\title{Map real property data to show vacant and improved properties in an area}
\usage{
map_vacancy(area = NULL)
}
\arguments{
\item{area}{sf class tibble. Object must include a name column.}
}
\description{
Map showing parcels described as principal residence, non-principal residence, vacant, and unimproved properties. If the area sf tibble includes multiple areas, a separate map is created for each area provided.
Parcel data is from the Maryland State Department of Assessment and Taxation.
}
