% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoning.R
\name{map_zoning}
\alias{map_zoning}
\title{Map zoning for a local area}
\usage{
map_zoning(area)
}
\arguments{
\item{area}{Required sf object with a 'name' column.}
}
\description{
Map zoning/zoning overlay codes for an area within the city.
The 2017 zoning data does not include any exemptions grants by the BMZA (Board of Municipal Zoning Appeals).
}
\examples{
\dontrun{
## Map zoning code for Bayview neighborhood
bayview <- get_area(area_type = "neighborhood", area_name = "Bayview")
map_zoning(area = bayview)
}
}
