% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbaltimore-package.R
\docType{package}
\name{mapbaltimore-package}
\alias{mapbaltimore}
\alias{mapbaltimore-package}
\title{mapbaltimore: Make maps for Baltimore City with open data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package provides data from the Baltimore City, the state of Maryland,
  and other sources, functions to access additional data, and function to create and modify
  simple maps of Baltimore neighborhoods using sf and ggplot2.
}
\keyword{internal}
