% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbaltimore.R
\docType{package}
\name{mapbaltimore}
\alias{mapbaltimore}
\title{mapbaltimore: A package for making maps with open data in Baltimore, Maryland.}
\description{
The mapbaltimore package is designed to provide an easy way to create maps of
Baltimore neighborhoods, Council districts, and other areas using open data on
schools, housing, parks, public transit, and more.
}
