% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maryland_open_data_api_key.R
\name{maryland_open_data_api_key}
\alias{maryland_open_data_api_key}
\title{Install a Maryland Open Data Portal API Key in Your \code{.Renviron} File for Repeated Use}
\usage{
maryland_open_data_api_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{The API key provided to you from Maryland Open Data Portal formatted in quotes. A key be be created after signing up \url{https://imap.maryland.gov/Pages/open-data-portal-signup.aspx}}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing MARYLAND_OPEN_DATA_API_KEY that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}
}
\description{
This function will add your Maryland Open Data Portal API key to your \code{.Renviron} file so it can be called securely without being stored
in your code. After you have installed your key, it can be called any time by typing \code{Sys.getenv("MARYLAND_OPEN_DATA_API_KEY")} and can be
used in package functions by simply typing MARYLAND_OPEN_DATA_API_KEY If you do not have an \code{.Renviron} file, the function will create on for you.
If you already have an \code{.Renviron} file, the function will append the key to your existing file, while making a backup of your
original file for disaster recovery purposes.
}
\examples{
\dontrun{
MARYLAND_OPEN_DATA_API_KEY("111111abc", install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("MARYLAND_OPEN_DATA_API_KEY")
}

\dontrun{
# If you need to overwrite an existing key:
MARYLAND_OPEN_DATA_API_KEY("111111abc", overwrite = TRUE, install = TRUE)
# First time, relead your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("MARYLAND_OPEN_DATA_API_KEY")
}
}
