% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_bus_lines}
\alias{mta_bus_lines}
\title{Maryland Transit Administration (MTA) Bus Routes (2022)}
\format{
A data frame with 103 rows and 4 variables:
\describe{
\item{\code{route_name}}{Bus route name}
\item{\code{route_type}}{Route type (CityLink, LocalLink, or Commuter Bus)}
\item{\code{route_number}}{Unique route number or color identifier}
\item{\code{route_abb}}{Route abbreviation (only different from \code{route_number}
for color CityLink routes)}
\item{\code{frequent}}{Logical indicator of route inclusion in MTA
BaltimoreLink's Frequent Transit Network.}
\item{\code{school}}{Indicator for school routes}
\item{\code{geometry}}{MULTILINESTRING bus route geometry}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/maryland-transit-mta-bus-lines-1}{Maryland Transit - MTA Bus Lines (MD iMap)}
}
\usage{
mta_bus_lines
}
\description{
Maryland Department of Transportation's Maryland Transit Administration
Summer 2022 Bus Routes including CityLink, LocalLink, Express BusLink and
Commuter Bus services and reflects bus route changes as of June 19, 2022. For
full details of service change visit:
\url{https://www.mta.maryland.gov/servicechanges/summer2022}
}
\keyword{datasets}
