% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_light_rail_lines}
\alias{mta_light_rail_lines}
\title{Maryland Transit Administration (MTA) Light RailLink Stations}
\format{
A data frame with 84 rows and 8 variables:
\describe{
\item{\code{id}}{Feature ID}
\item{\code{rail_name}}{Line name (Light Rail Line)}
\item{\code{mode}}{Facility mode (Light Rail)}
\item{\code{tunnel}}{Tunnel indicator}
\item{\code{direction}}{Travel direction}
\item{\code{miles}}{Section mileage}
\item{\code{status}}{Section status}
\item{\code{geometry}}{LINESTRING line geometry}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/maryland-transit-light-rail-lines/}{Maryland Transit - Light Rail Lines (MD iMap)}
}
\usage{
mta_light_rail_lines
}
\description{
Location of MTA Light Rail Stations.
}
\keyword{datasets}
