% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_light_rail_stations}
\alias{mta_light_rail_stations}
\title{Maryland Transit Administration (MTA) Light RailLink Stations}
\format{
A data frame with 33 rows and 11 variables:
\describe{
\item{\code{id}}{integer COLUMN_DESCRIPTION}
\item{\code{name}}{character COLUMN_DESCRIPTION}
\item{\code{address}}{character COLUMN_DESCRIPTION}
\item{\code{city}}{character COLUMN_DESCRIPTION}
\item{\code{state}}{character COLUMN_DESCRIPTION}
\item{\code{zipcode}}{integer COLUMN_DESCRIPTION}
\item{\code{mode}}{character COLUMN_DESCRIPTION}
\item{\code{avg_wkdy}}{integer COLUMN_DESCRIPTION}
\item{\code{avg_wknd}}{character COLUMN_DESCRIPTION}
\item{\code{facility_type}}{character COLUMN_DESCRIPTION}
\item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/maryland-transit-light-raillink-stations/}{Maryland Transit - Light RailLink Stations (MD iMap)}
}
\usage{
mta_light_rail_stations
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
