% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_marc_lines}
\alias{mta_marc_lines}
\title{MTA MARC Train Lines}
\format{
A data frame with 162 rows and 8 variables:
\describe{
  \item{\code{id}}{integer COLUMN_DESCRIPTION}
  \item{\code{rail_name}}{character COLUMN_DESCRIPTION}
  \item{\code{mode}}{character COLUMN_DESCRIPTION}
  \item{\code{tunnel}}{character COLUMN_DESCRIPTION}
  \item{\code{direction}}{character COLUMN_DESCRIPTION}
  \item{\code{miles}}{double COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
  \item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/de0efbe9f8884ac5aa69864b6b3ff633_10/}{Maryland Transit - MARC Train Lines (MD iMap)}
}
\usage{
mta_marc_lines
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
