% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_marc_stations}
\alias{mta_marc_stations}
\title{MTA MARC Train Stations}
\format{
A data frame with 44 rows and 12 variables:
\describe{
  \item{\code{id}}{integer COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{address}}{character COLUMN_DESCRIPTION}
  \item{\code{city}}{character COLUMN_DESCRIPTION}
  \item{\code{state}}{character COLUMN_DESCRIPTION}
  \item{\code{zipcode}}{integer COLUMN_DESCRIPTION}
  \item{\code{line_name}}{character COLUMN_DESCRIPTION}
  \item{\code{mode}}{character COLUMN_DESCRIPTION}
  \item{\code{avg_wkdy}}{integer COLUMN_DESCRIPTION}
  \item{\code{avg_wknd}}{integer COLUMN_DESCRIPTION}
  \item{\code{facility_type}}{character COLUMN_DESCRIPTION}
  \item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/maryland-transit-marc-trains-stations/}{Maryland Transit - MARC Trains Stations (MD iMap)}
}
\usage{
mta_marc_stations
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
