% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_subway_lines}
\alias{mta_subway_lines}
\title{Maryland Transit Administration (MTA) SubwayLink Metro Lines}
\format{
A data frame with 34 rows and 8 variables:
\describe{
  \item{\code{id}}{integer COLUMN_DESCRIPTION}
  \item{\code{rail_name}}{character COLUMN_DESCRIPTION}
  \item{\code{mode}}{character COLUMN_DESCRIPTION}
  \item{\code{tunnel}}{character COLUMN_DESCRIPTION}
  \item{\code{direction}}{character COLUMN_DESCRIPTION}
  \item{\code{miles}}{double COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
  \item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\usage{
mta_subway_lines
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
