% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_subway_stations}
\alias{mta_subway_stations}
\title{Maryland Transit Administration (MTA) SubwayLink Metro Stations}
\format{
A data frame with 14 rows and 10 variables:
\describe{
\item{\code{id}}{Station identification number as integer}
\item{\code{name}}{Station name}
\item{\code{address}}{Station street address}
\item{\code{city}}{City}
\item{\code{state}}{State}
\item{\code{mode}}{Travel mode (Metro)}
\item{\code{avg_wkdy}}{Average weekday passengers}
\item{\code{avg_wknd}}{Average weekend passengers}
\item{\code{facility_type}}{Facility type (Station)}
\item{\code{geometry}}{POINT station location geometry}
}
}
\source{
\href{https://geodata.md.gov/imap/rest/services/Transportation/MD_Transit/FeatureServer/4}{Baltimore Metro SubwayLink Stations}
}
\usage{
mta_subway_stations
}
\description{
Location of MTA SubwayLink Metro Stations.
}
\keyword{datasets}
