% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndc_projects}
\alias{ndc_projects}
\title{Neighborhood Design Center Projects}
\format{
A data frame with 2487 rows and 16 variables:
\describe{
\item{\code{ndc_office}}{character NDC Office}
\item{\code{ndc_program}}{character NDC Program}
\item{\code{project_id}}{double Project ID}
\item{\code{project_name}}{character Project name}
\item{\code{project_type}}{character Project type}
\item{\code{project_status}}{character Project status}
\item{\code{start_date}}{character Start date}
\item{\code{end_date}}{character End date}
\item{\code{client_name}}{character Client name}
\item{\code{address}}{character Street address}
\item{\code{city}}{character City}
\item{\code{state}}{character State}
\item{\code{zip_code}}{integer Zip code}
\item{\code{geometry}}{list Geometry}
\item{\code{start_year}}{character Start year}
\item{\code{end_year}}{character End year}
}
}
\usage{
ndc_projects
}
\description{
A reference to project completed by the Neighborhood Design
Center through 2022.
}
\details{
DETAILS
}
\keyword{internal}
