% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhoods}
\alias{neighborhoods}
\title{Neighborhood boundaries for Baltimore City}
\format{
A data frame with 278 rows and 2 variables:
\describe{
\item{\code{name}}{Neighborhood name}
\item{\code{type}}{Type of area, with options including residential, industrial area, park/open space, institutionl area and business park)}
\item{\code{acres}}{Area of the neighborhood (acres)}
\item{\code{osm_id}}{Open Street Map (OSM) relation identifier}
\item{\code{wikidata}}{Wikidata entity identifier}
\item{\code{geometry}}{MULITPOLYGON boundary geometry}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/fc5d183b20a145009eae8f8b171eeb0d_0}{Maryland Baltimore City Neighborhoods (MD iMap)}
}
\usage{
neighborhoods
}
\description{
Baltimore City Neighborhoods or Neighborhood Statistical Areas
}
\keyword{datasets}
