% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhoods_2020}
\alias{neighborhoods_2020}
\title{Neighborhood Boundaries for Baltimore City (2020)}
\format{
A data frame with 279 rows and 8 variables:
\describe{
\item{\code{name}}{Neighborhood name}
\item{\code{name_alt}}{2010 neighborhood name}
\item{\code{type}}{Type of area, with options including residential, industrial area, park/open space, institutionl area and business park)}
\item{\code{acres}}{Area of the neighborhood (acres)}
\item{\code{osm_id}}{Open Street Map (OSM) relation identifier}
\item{\code{wikidata}}{Wikidata entity identifier}
\item{\code{color_id}}{Color identifier}
\item{\code{geometry}}{MULITPOLYGON boundary geometry}
}
}
\source{
\href{https://services1.arcgis.com/43Lm3JYE3nM91DAF/arcgis/rest/services/NSA_March2023/FeatureServer/0}{NSA_Feb2023_service}
}
\usage{
neighborhoods_2020
}
\description{
Baltimore City neighborhoods (officially known as Neighborhood Statistical
Areas) established by the Baltimore City Department of Planning based on the
2020 U.S. Decennial Census. This is an updated version of the 2010
Neighborhood Statistical Areas.
}
\keyword{datasets}
