% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{park_districts}
\alias{park_districts}
\title{Baltimore Park Districts}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{\code{name}}{Park district name}
\item{\code{geometry}}{MULTIPOLYGON geometry with park district boundaries}
}
}
\usage{
park_districts
}
\description{
Park districts for the \href{https://bcrp.baltimorecity.gov/}{Baltimore City Department of Recreation and Parks}. District boundaries are used for
park maintenance administration.
}
\keyword{datasets}
