% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parks}
\alias{parks}
\title{Baltimore City Parks}
\format{
A data frame with 297 rows and 6 variables:
\describe{
  \item{park_id}{...}
  \item{name}{...}
  \item{address}{...}
  \item{name_alt}{...}
  \item{bcrp}{...}
  \item{geoms}{...}
  ...
}
}
\source{
\url{...}
}
\usage{
parks
}
\description{
...
}
\keyword{datasets}
