% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planning_districts}
\alias{planning_districts}
\title{Baltimore City Planning Districts}
\format{
A data frame with 11 rows and 3 variables:
\describe{
  \item{name}{...}
  \item{abb}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{...}
}
\usage{
planning_districts
}
\description{
...
}
\keyword{datasets}
