% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_open_baltmore_data.R
\name{plot_permits}
\alias{plot_permits}
\title{Plot selected housing permits from Open Baltimore data portal}
\usage{
plot_permits(permits, area_label, plot_by = c("month", "quarter", "year"))
}
\arguments{
\item{permits}{A data frame of building permit data.}

\item{area_label}{Required character vector used as label for map to describe the geographic area covered by the permit data.}

\item{plot_by}{Required character vector to select the period of time to aggregate permit count. Supports "month", "quarter", and "year"}
}
\description{
This function plots a basic chart of building permits issued by month, quarter, or year. Expected to use in combination with the \code{get_permits} function.
}
