% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police_districts_2023}
\alias{police_districts_2023}
\title{Baltimore City Police Districts (2023-Current)}
\format{
A data frame with 9 rows and 4 variables:
\describe{
\item{\code{id}}{Police district number}
\item{\code{name}}{Police district name}
\item{\code{name_abb}}{District name abbreviation}
\item{\code{geometry}}{MULTIPOLYGON geometry for district boundary}
}
}
\source{
\url{https://services1.arcgis.com/UWYHeuuJISiGmgXx/arcgis/rest/services/Police_District/FeatureServer/0}
}
\usage{
police_districts_2023
}
\description{
Baltimore City Police Districts boundaries updated in 2023.
}
\keyword{datasets}
