% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rec_centers}
\alias{rec_centers}
\title{Baltimore City Recreation Centers}
\format{
A data frame with 48 rows and 18 variables:
\describe{
\item{\code{id}}{ID}
\item{\code{name}}{Center name}
\item{\code{name_short}}{Short name}
\item{\code{street_address}}{Street address}
\item{\code{address}}{Full address}
\item{\code{center_amenities}}{Center amenities}
\item{\code{center_assets}}{Center facility assets}
\item{\code{center_category}}{Center category}
\item{\code{center_type}}{Center type}
\item{\code{school_name}}{School name (school-based centers only)}
\item{\code{operator}}{Operator (BCRP only)}
\item{\code{recreation_district}}{BCRP Recreation district}
\item{\code{park_district}}{BCRP Park Maintenance District}
\item{\code{council_district}}{City Council district}
\item{\code{legislative_district}}{Maryland legislative district}
\item{\code{police_district}}{BPD Police district (maybe outdated)}
\item{\code{notes}}{Notes}
\item{\code{geometry}}{POINT geometry with center location}
}
}
\source{
\url{https://services1.arcgis.com/UWYHeuuJISiGmgXx/arcgis/rest/services/recreationCenter2023/FeatureServer}
}
\usage{
rec_centers
}
\description{
Currently includes only publicly operated (BCRP) rec centers. Expect to add
private operator facilities. Added 2023-10-19.
}
\keyword{datasets}
