% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{request_types}
\alias{request_types}
\title{311 Service Request Types for Baltimore City}
\format{
A data frame with 320 rows and 1 variables:
\describe{
  \item{request_type}{...}
  ...
}
}
\source{
\url{...}
}
\usage{
request_types
}
\description{
A list of request types based on unique request types used between January 2019 and October 2020.
}
\keyword{datasets}
