% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{respagency_codes}
\alias{respagency_codes}
\title{Baltimore City Real Property Responsible Agency Codes}
\format{
A data frame with 37 rows and 7 variables:
\describe{
\item{\code{name}}{Responsible agency name}
\item{\code{code}}{Responsible agency code}
\item{\code{agency_name}}{Baltimore City agency/commission name}
\item{\code{agency_abb}}{Baltimore City agency/commission abbreviation}
\item{\code{division_name}}{Agency division name}
\item{\code{active_code}}{Active code indicator (\code{FALSE} for codes that do not appear in data)}
\item{\code{notes}}{Notes on code/agency}
}
}
\source{
\url{https://docs.google.com/spreadsheets/d/1Dnyp4-AZxvFPpt5Vci4NRWR9tGP99R8RaHuPCbzcGCA/edit?usp=sharing}
}
\usage{
respagency_codes
}
\description{
A reference table of responsible agency codes appearing in the Baltimore City
real property data used by \code{\link[=get_area_property]{get_area_property()}}. Updated 2023 March 29.
}
\keyword{datasets}
