% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schools_21stc}
\alias{schools_21stc}
\title{Baltimore 21st Century Schools}
\format{
A data frame with 29 rows and 24 variables:
\describe{
\item{\code{school_name}}{character COLUMN_DESCRIPTION}
\item{\code{school_number}}{double COLUMN_DESCRIPTION}
\item{\code{nces_number}}{double COLUMN_DESCRIPTION}
\item{\code{grade_band}}{character COLUMN_DESCRIPTION}
\item{\code{url}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{double COLUMN_DESCRIPTION}
\item{\code{type}}{character COLUMN_DESCRIPTION}
\item{\code{bldg_budget_approx}}{double COLUMN_DESCRIPTION}
\item{\code{status_21c}}{character COLUMN_DESCRIPTION}
\item{\code{status_inspire}}{character COLUMN_DESCRIPTION}
\item{\code{inspire_plan}}{character COLUMN_DESCRIPTION}
\item{\code{occupancy_month}}{double COLUMN_DESCRIPTION}
\item{\code{occupancy_year}}{double COLUMN_DESCRIPTION}
\item{\code{address}}{character COLUMN_DESCRIPTION}
\item{\code{city}}{character COLUMN_DESCRIPTION}
\item{\code{state}}{character COLUMN_DESCRIPTION}
\item{\code{zip}}{double COLUMN_DESCRIPTION}
\item{\code{phone}}{double COLUMN_DESCRIPTION}
\item{\code{alt_school_name}}{character COLUMN_DESCRIPTION}
\item{\code{bldg_name}}{character COLUMN_DESCRIPTION}
\item{\code{alt_name}}{character COLUMN_DESCRIPTION}
\item{\code{lon}}{double COLUMN_DESCRIPTION}
\item{\code{lat}}{double COLUMN_DESCRIPTION}
\item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\usage{
schools_21stc
}
\description{
Schools with buildings in the 21st Century Schools Program
}
\details{
Note: this documentation is a placeholder.

DETAILS
}
\keyword{datasets}
