% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_map_limits.R
\name{set_map_limits}
\alias{set_map_limits}
\title{Set map limits to area}
\usage{
set_map_limits(
  area = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  expand = FALSE,
  ...
)
}
\arguments{
\item{area}{\code{sf} object to buffer and/or adjust.}

\item{bbox}{\code{bbox} object to buffer and/or adjust. If an area is provided,
any bbox is ignored.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crs}{Coordinate reference system to use for \code{coord_sf()}. Default
\code{pkgconfig::get_config("mapbaltimore.crs", 2804)}}

\item{expand}{Default FALSE. If TRUE, use scale_y_continuous and
scale_x_continuous to expand map extent to provided parameters.}

\item{...}{Additional parameters to pass to \code{coord_sf()}.}
}
\value{
\code{ggplot2::coord_sf()} function with xlim and ylim parameters
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because the functionality has been incorporated
into the improved \code{\link[maplayer:layer_neatline]{maplayer::layer_neatline()}} function which uses a similar
set of parameters.

Set limits for a map to the bounding box of an area using \code{coord_sf()}.
Optionally, adjust the area size by applying a buffer and/or adjust the
aspect ratio of the limiting bounding box to match a set aspect ratio.
}
\details{
Set map limits to area with optional buffer or aspect ratio adjustment
}
\examples{
# Show detailed city boundary with map focused on area of Fell's Point (with 50m buffer)
library(ggplot2)

ggplot() +
  geom_sf(data = baltimore_city_detailed) +
  set_map_limits(area = get_area("neighborhood", "Fells Point"), dist = 50)
}
\keyword{internal}
