% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_map_theme}
\alias{set_map_theme}
\title{Set default map theme}
\usage{
set_map_theme(map_theme = NULL, show_axis = FALSE)
}
\arguments{
\item{map_theme}{ggplot2 theme. Optional. Defaults to \code{\link[ggplot2]{theme_minimal()}}}

\item{show_axis}{Logical. If TRUE, keep theme axis formatting. If FALSE, hide the panel grid, axis title, and axis text.}
}
\description{
Set a map theme using \code{\link[ggplot2]{theme_set()}} and default for \code{geom_label} using \code{\link[ggplot2]{update_geom_defaults()}}. Optionally hides axis text and labels.
}
