% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streets}
\alias{streets}
\title{Baltimore City Street Center lines}
\format{
Simple feature collection with 48,473 features and 23 fields.
\describe{
  \item{type}{...}
  \item{subtype}{...}
  \item{subtype_label}{...}
  \item{dirpre}{...}
  \item{feanme}{...}
  \item{featype}{...}
  \item{dirsuf}{...}
  \item{fraddl}{...}
  \item{toaddl}{...}
  \item{fraddr}{...}
  \item{toaddr}{...}
  \item{fraddla}{...}
  \item{toaddla}{...}
  \item{fraddra}{...}
  \item{toaddra}{...}
  \item{leftzip}{...}
  \item{rightzip}{...}
  \item{fullname}{...}
  \item{sha_class}{...}
  \item{sha_class_label}{...}
  \item{blocktext}{...}
  \item{block_num}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{https://dotgis.baltimorecity.gov/arcgis/rest/services/DOT_Map_Services/DOT_Basemap/MapServer/7}
}
\usage{
streets
}
\description{
...
}
\keyword{datasets}
