% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wards_1797_1918}
\alias{wards_1797_1918}
\title{Historic Ward Boundaries, 1797-1918 for Baltimore City}
\format{
A data frame with 245 rows and 4 variables:
\describe{
  \item{\code{year}}{Earliest effective year of ward boundary}
  \item{\code{name}}{Ward name}
  \item{\code{number}}{Ward number}
  \item{\code{geometry}}{Multipolygons with ward boundary for year}
}
}
\source{
\url{https://msa.maryland.gov/bca/wards/index.html}
}
\usage{
wards_1797_1918
}
\description{
Historic ward boundary data from 1797 to 1918. Derived from KML data provided by the Baltimore City Archives.
}
\keyword{datasets}
