% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk_block2tract}
\alias{xwalk_block2tract}
\title{U.S. Census Block-to-Tract Crosswalk with 2010 Block Household Population}
\format{
A data frame with 13598 rows and 3 variables:
\describe{
\item{\code{block}}{Block GeoID}
\item{\code{tract}}{Tract GeoID}
\item{\code{households}}{Block household population}
}
}
\usage{
xwalk_block2tract
}
\description{
Used to generate xwalk_neighborhood2tract.
}
\keyword{datasets}
