% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk_neighborhood2tract}
\alias{xwalk_neighborhood2tract}
\title{Neighborhood-to-U.S. Census Tract Crosswalk}
\format{
A data frame with 551 rows and 4 variables:
\describe{
  \item{\code{name}}{Neighborhood name}
  \item{\code{geoid}}{GeoID for U.S. Census tract}
  \item{\code{tract}}{Tract number}
  \item{\code{weight}}{Share of total households in neighborhood and U.S. Census tract}
}
}
\usage{
xwalk_neighborhood2tract
}
\description{
Share of total households is based on the proportion of U.S.
  Census tract population within the named neighborhood based on overlapping
  U.S. Census Block groups.
}
\keyword{datasets}
