% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk_zip2csa}
\alias{xwalk_zip2csa}
\title{Zipcode-to-Community Statistical Area (NSA) Crosswalk}
\format{
A data frame with 119 rows and 3 variables:
\describe{
\item{\code{zip}}{Zipcode}
\item{\code{csa}}{Community Statistical Area name}
\item{\code{id}}{Community Statistical Area id number}
}
}
\source{
\url{https://bniajfi.org/mapping-resources/}
}
\usage{
xwalk_zip2csa
}
\description{
A crosswalk to match zipcodes to Community Statistical Areas.
}
\keyword{datasets}
