% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoning}
\alias{zoning}
\title{Baltimore City Zoning Code}
\format{
A data frame with 2,406 rows and 4 variables:
\describe{
\item{\code{zoning}}{Zoning designation code}
\item{\code{overlay}}{Overlay zone designation}
\item{\code{label}}{Label combining zoning and overlay zoning codes}
\item{\code{category_zoning}}{Zoning code category}
\item{\code{name_zoning}}{Zoning code name}
\item{\code{category_overlay}}{Overlay code category}
\item{\code{name_overlay}}{Overlay zoning name}
\item{\code{geometry}}{MULTIPOLYGON geometry for zoning areas}
}
}
\source{
\url{https://geodata.baltimorecity.gov/egis/rest/services/Planning/Boundaries_and_Plans/MapServer/20}
}
\usage{
zoning
}
\description{
The Baltimore City Zoning Code is administered by the Baltimore City
Department of Housing and Community Development (HCD) Office of the Zoning
Administrator. This office supports the Board of Municipal Zoning Appeals
(BMZA).
}
\keyword{datasets}
