% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adopted_plans}
\alias{adopted_plans}
\title{Adopted city plans, accepted community-initiated plans, and LINCS corridors}
\format{
A data frame with 58 rows and 5 variables:
\describe{
  \item{name}{Plan or area name}
  \item{year_adopted}{Year adopted or initiated}
  \item{program}{Planning program}
  \item{url}{URL to website or planning document}
  \item{geometry}{multipolygon for plans and multilinestring for LINCS corridors}
}
}
\source{
\url{...}
}
\usage{
adopted_plans
}
\description{
Adopted city plans, accepted community-initiated plans, and LINCS corridors
}
\keyword{datasets}
