% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adopted_plans}
\alias{adopted_plans}
\title{Adopted city plans, accepted community-initiated plans, and LINCS corridors}
\format{
A data frame with 58 rows and 5 variables:
\describe{
  \item{\code{plan_name}}{Plan or area name}
  \item{\code{year_adopted}}{Year adopted or initiated}
  \item{\code{program}}{Planning program}
  \item{\code{url}}{URL of plan website or document}
  \item{\code{geometry}}{Multipolygon for plan areas and multilinestring for LINCS corridors}
}
}
\source{
\url{...}
}
\usage{
adopted_plans
}
\description{
Combined area plans and LINCS corridor data from the Baltimore City
Department of Planning.
}
\keyword{datasets}
