% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{balt_tbl_labs}
\alias{balt_tbl_labs}
\title{Baltimore data table labels}
\format{
A data frame with 9 rows and 7 variables:
\describe{
\item{\code{fn_name}}{character Function name}
\item{\code{table}}{character Table name}
\item{\code{col}}{character Column name}
\item{\code{label}}{character Column label}
\item{\code{definition}}{logical Column variable definition}
\item{\code{source}}{logical Column variable data source}
\item{\code{fmt}}{character Column data format}
}
}
\usage{
balt_tbl_labs
}
\description{
From \href{https://docs.google.com/spreadsheets/d/1FXEJlhccnhoQmSO2WydBidXIw-f2lpomURDGy9KBgJw/edit?usp=sharing}{Google Sheet}
}
\keyword{datasets}
