% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_bbox}
\alias{baltimore_bbox}
\title{Baltimore City bounding box in EPSG:4326 CRS}
\format{
A bbox class object.
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_bbox
}
\description{
A generalized boundary for Baltimore City, Maryland converted to a bounding
box object.
}
\keyword{datasets}
