% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_block_groups}
\alias{baltimore_block_groups}
\title{U.S. Census Block Groups in Baltimore City}
\format{
A data frame with 653 rows and 9 variables:
\describe{
  \item{tractce}{...}
  \item{blkgrpce}{...}
  \item{geoid}{...}
  \item{namelsad}{...}
  \item{aland}{...}
  \item{awater}{...}
  \item{intptlat}{...}
  \item{intptlon}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_block_groups
}
\description{
...
}
\keyword{datasets}
