% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_block_groups}
\alias{baltimore_block_groups}
\title{U.S. Census Block Groups in Baltimore City}
\format{
A data frame with 653 rows and 9 variables:
\describe{
  \item{\code{tractce}}{Current census tract code}
  \item{\code{blkgrpce}}{Current block group number}
  \item{\code{geoid}}{Census block group identifier; a concatenation of the current state FIPS code, county FIPS code, census tract code, and block group number}
  \item{\code{namelsad}}{Current translated legal/statistical area description and the block group number}
  \item{\code{aland}}{Current land area (square meters)}
  \item{\code{awater}}{Current water area (square meters)}
  \item{\code{intptlat}}{Current latitude of the internal point}
  \item{\code{intptlon}}{Current longitude of the internal point}
  \item{\code{geometry}}{Polygon with block group boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_block_groups
}
\description{
...
}
\keyword{datasets}
