% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_block_groups}
\alias{baltimore_block_groups}
\title{U.S. Census Block Groups}
\format{
A data frame with 653 rows and 9 variables:
\describe{
  \item{\code{tractce}}{census tract code}
  \item{\code{blkgrpce}}{block group number}
  \item{\code{geoid}}{Census block group identifier; a concatenation of the state FIPS code, county FIPS code, census tract code, and block group number}
  \item{\code{namelsad}}{translated legal/statistical area description and the block group number}
  \item{\code{aland}}{land area (square meters)}
  \item{\code{awater}}{water area (square meters)}
  \item{\code{intptlat}}{latitude of the internal point}
  \item{\code{intptlon}}{longitude of the internal point}
  \item{\code{geometry}}{Polygon with block group boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_block_groups
}
\description{
...
}
\keyword{datasets}
