% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_block_groups}
\alias{baltimore_block_groups}
\title{U.S. Census Block Groups - 2020}
\format{
A data frame with 618 rows and 9 variables:
\describe{
\item{\code{tractce}}{Census tract code}
\item{\code{blkgrpce}}{Census block group number}
\item{\code{geoid}}{Census block group identifier; a concatenation of the state FIPS code, county FIPS code, census tract code, and block group number}
\item{\code{namelsad}}{translated legal/statistical area description and the block group number}
\item{\code{aland}}{land area (square meters)}
\item{\code{awater}}{water area (square meters)}
\item{\code{intptlat}}{latitude of the internal point}
\item{\code{intptlon}}{longitude of the internal point}
\item{\code{geometry}}{POLYGON geometry for block group boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_block_groups
}
\description{
U.S. Census Block Groups for Baltimore city, Maryland downloaded from the
U.S. Census Bureau API with the tigris package.
}
\keyword{datasets}
