% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_blocks}
\alias{baltimore_blocks}
\title{U.S. Census Blocks (2010 Decennial) in Baltimore City}
\format{
A data frame with 13,598 rows and 9 variables:
\describe{
  \item{tractce10}{Tract FIPS}
  \item{blkgrpce10}{Block group FIPS}
  \item{geoid10}{Block GeoID}
  \item{name10}{Block name}
  \item{aland10}{Land area}
  \item{awater10}{Water area}
  \item{intptlat10}{Interior center point latitude}
  \item{intptlon10}{Interior center point longitude}
  \item{geometry}{Multipolygon with block boundary}
  ...
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_blocks
}
\description{
...
}
\keyword{datasets}
