% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltimorecensus.R
\docType{data}
\name{baltimore_blocks}
\alias{baltimore_blocks}
\title{U.S. Census Blocks - 2020}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 13598 rows and 9 columns.
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_blocks
}
\description{
U.S. Census Blocks for Baltimore city, Maryland downloaded from the U.S.
Census Bureau API with the tigris package.
}
\keyword{datasets}
