% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_city}
\alias{baltimore_city}
\title{Generalized political boundary for Baltimore City}
\format{
A data frame with 1 row and 3 variables:
\describe{
  \item{name}{County name}
  \item{countyfp}{Unique county FIPS code}
  \item{geoid}{County GeoID}
  \item{aland}{Land area}
  \item{awater}{Water area}
  \item{intptlat}{Interior center point latitude}
  \item{intptlon}{Interior center point longitude}
  \item{geometry}{Multipolygon with the boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_city
}
\description{
A generalized boundary for Baltimore City, Maryland
from statewide dataset of generalized county boundaries.
}
\keyword{datasets}
