% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_city}
\alias{baltimore_city}
\title{A generalized political boundary for Baltimore City}
\format{
A data frame with 1 row and 3 variables:
\describe{
  \item{county}{name of the county}
  \item{county_fips}{Unique county FIPS code}
  \item{geometry}{multipolygon with the boundary}
  ...
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/maryland-physical-boundaries-county-boundaries-generalized}
}
\usage{
baltimore_city
}
\description{
A generalized boundary for Baltimore City, Maryland
from statewide dataset of generalized county boundaries.
}
\keyword{datasets}
