% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_city_detailed}
\alias{baltimore_city_detailed}
\title{Detailed physical boundary for Baltimore City}
\format{
A data frame with 1 row and 3 variables:
\describe{
\item{\code{name}}{County name}
\item{\code{countyfp}}{3-character county FIPS code}
\item{\code{geometry}}{MULITPOLYGON boundary geometry}
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/maryland-physical-boundaries-county-boundaries-detailed}
}
\usage{
baltimore_city_detailed
}
\description{
A detailed physical boundary of Baltimore City.
}
\keyword{datasets}
