% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_city_detailed}
\alias{baltimore_city_detailed}
\title{A detailed physical boundary for Baltimore City}
\format{
A data frame with 1 row and 3 variables:
\describe{
  \item{name}{name of the county}
  \item{countyfp}{Unique county FIPS code}
  \item{geometry}{multipolygon with the boundary}
}
}
\source{
\url{}
}
\usage{
baltimore_city_detailed
}
\description{
...
}
\keyword{datasets}
