% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_gis_index}
\alias{baltimore_gis_index}
\title{Baltimore ArcGIS Server index data}
\format{
A data frame with 1,286 rows and 17 variables:
\describe{
\item{\code{name}}{Name}
\item{\code{nm}}{Name with snake case}
\item{\code{type}}{Service/layer type}
\item{\code{url}}{Folder/service/layer URL}
\item{\code{urlType}}{URL type}
\item{\code{folderPath}}{Index type}
\item{\code{serviceName}}{Service name}
\item{\code{serviceType}}{Service type}
\item{\code{id}}{integer Layer ID number}
\item{\code{parentLayerId}}{integer Parent layer ID number}
\item{\code{defaultVisibility}}{logical Layer default visibility}
\item{\code{minScale}}{double Minimum scale}
\item{\code{maxScale}}{integer Maximum scale}
\item{\code{geometryType}}{Geometry type}
\item{\code{subLayerIds}}{list Sublayer ID numbers}
\item{\code{supportsDynamicLegends}}{logical Supports dynamic legends}
}
}
\usage{
baltimore_gis_index
}
\description{
A data.frame indexing the layers, services, and folders on four ArcGIS
Servers maintained by the Baltimore City Mayor's Office of Information
Technology (MOIT) Enterprise GIS (EGIS) program. A limited number of
potential sensitive and unresponsive server layers have been excluded. Used
by the \code{\link[=get_baltimore_data]{get_baltimore_data()}} function. Updated December 23, 2022.
}
\keyword{datasets}
