% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_gis_index}
\alias{baltimore_gis_index}
\title{Baltimore ArcGIS Server index data}
\format{
A data frame with 1021 rows and 15 variables:
\describe{
\item{\code{name}}{character Name}
\item{\code{nm}}{character Name with snake case}
\item{\code{index}}{character Index type}
\item{\code{type}}{character Service/layer type}
\item{\code{server_type}}{character Server type}
\item{\code{url}}{character Folder/service/layer URL}
\item{\code{parent}}{character Parent folder/service/layer}
\item{\code{id}}{integer Layer ID number}
\item{\code{parentLayerId}}{integer Parent layer ID number}
\item{\code{defaultVisibility}}{logical Layer default visibility}
\item{\code{minScale}}{double Minimum scale}
\item{\code{maxScale}}{integer Maximum scale}
\item{\code{geometryType}}{character Geometry type}
\item{\code{subLayerIds}}{list Sublayer ID numbers}
\item{\code{supportsDynamicLegends}}{logical Supports dynamic legends}
}
}
\usage{
baltimore_gis_index
}
\description{
A dataframe indexing the layers, services, and folders on two ArcGIS Servers
maintained by the Baltimore City Mayor's Office of Information Technology
(MOIT) Enterprise GIS (EGIS) program. Used by the \code{\link[=get_baltimore_data]{get_baltimore_data()}}
function. Updated November 14, 2022.
}
\keyword{datasets}
