% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_mihp}
\alias{baltimore_mihp}
\title{Maryland Inventory of Historic Properties in Baltimore City}
\format{
A data frame with 5,203 rows and 14 variables:
\describe{
  \item{\code{num_polys}}{...}
  \item{\code{mihp_id}}{...}
  \item{\code{property_id}}{...}
  \item{\code{mihp_num}}{...}
  \item{\code{name}}{...}
  \item{\code{alternate_name}}{...}
  \item{\code{full_address}}{...}
  \item{\code{town}}{...}
  \item{\code{county}}{...}
  \item{\code{pdflink}}{...}
  \item{\code{xcoord}}{...}
  \item{\code{ycoord}}{...}
  \item{\code{do_erecord}}{...}
  \item{\code{geoms}}{...}
}
}
\source{
\url{...}
}
\usage{
baltimore_mihp
}
\description{
...
}
\keyword{datasets}
