% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltimorecensus.R
\docType{data}
\name{baltimore_msa_counties}
\alias{baltimore_msa_counties}
\title{County boundaries for the Baltimore–Columbia–Towson MSA}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7 rows and 13 columns.
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_msa_counties
}
\description{
Counties boundaries in the Baltimore–Columbia–Towson Metropolitan Statistical Area (MSA)
include Baltimore City, Baltimore County, Carroll County, Anne Arundel County,
Howard County, Queen Anne's County, and Harford County.
}
\keyword{datasets}
