% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_msa_counties}
\alias{baltimore_msa_counties}
\title{County boundaries for the Baltimore–Columbia–Towson MSA}
\format{
A data frame with 7 rows and 18 variables:
\describe{
  \item{\code{statefp}}{State FIPS code for Maryland}
  \item{\code{countyfp}}{3-character county FIPS code}
  \item{\code{countyns}}{..}
  \item{\code{geoid}}{Unique county FIPS code}
  \item{\code{name}}{County name}
  \item{\code{namelsad}}{concatenated variable length geographic area name and legal/statistical area description (LSAD)}
  \item{\code{lsad}}{Legal/statistical area description (LSAD)}
  \item{\code{classfp}}{character}
  \item{\code{mtfcc}}{5-digit MAF/TIGER Feature Class Code (MTFCC)}
  \item{\code{csafp}}{character}
  \item{\code{cbsafp}}{character}
  \item{\code{metdivfp}}{...}
  \item{\code{funcstat}}{Current functional status}
  \item{\code{aland}}{Current land area (square meters)}
  \item{\code{awater}}{Current water area (square meters)}
  \item{\code{intptlat}}{Current latitude of the internal point}
  \item{\code{intptlon}}{Current longitude of the internal point}
  \item{\code{geometry}}{Multipolygon with the county boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_msa_counties
}
\description{
Counties boundaries in the Baltimore–Columbia–Towson Metropolitan Statistical Area (MSA)
include Baltimore City, Baltimore County, Carroll County, Anne Arundel County,
Howard County, Queen Anne's County, and Harford County.
}
\keyword{datasets}
