% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_msa_counties}
\alias{baltimore_msa_counties}
\title{County boundaries for the Baltimore–Columbia–Towson MSA}
\format{
A data frame with 7 rows and 18 variables:
\describe{
  \item{statefp}{State FIPS code for Maryland}
  \item{countyfp}{Unique county FIPS code}
  \item{countyns}{..}
  \item{geoid}{Unique county FIPS code}
  \item{name}{County name}
  \item{namelsad}{...}
  \item{lsad}{...}
  \item{csafp}{...}
  \item{metdivfp}{...}
  \item{funcstat}{...}
  \item{aland}{Land area}
  \item{awater}{Water area}
  \item{intptlat}{Interior center point latitude}
  \item{intptlat}{Interior center point longitude}
  \item{geometry}{Multipolygon with the county boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_msa_counties
}
\description{
Counties boundaries in the Baltimore–Columbia–Towson Metropolitan Statistical Area (MSA)
include Baltimore City, Baltimore County, Carroll County, Anne Arundel County,
Howard County, Queen Anne's County, and Harford County.
}
\keyword{datasets}
