% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltimorecensus.R
\docType{data}
\name{baltimore_tracts}
\alias{baltimore_tracts}
\title{U.S. Census Tracts - 2020}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 200 rows and 9 columns.
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_tracts
}
\description{
U.S. Census Tracts for Baltimore city, Maryland downloaded from the
U.S. Census Bureau API with the tigris package.
}
\keyword{datasets}
