% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_tracts}
\alias{baltimore_tracts}
\title{U.S. Census Tracts in Baltimore City}
\format{
A data frame with 200 rows and 9 variables:
\describe{
  \item{tractce}{...}
  \item{geoid}{Tract GeoID}
  \item{name}{...}
  \item{namelsad}{...}
  \item{aland}{Land area}
  \item{awater}{Water area}
  \item{intptlat}{Interior center point latitude}
  \item{intptlon}{Interior center point longitude}
  \item{geometry}{Polygon with tract boundary}
  ...
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_tracts
}
\description{
...
}
\keyword{datasets}
