% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_tracts}
\alias{baltimore_tracts}
\title{U.S. Census Tracts}
\format{
A data frame with 200 rows and 9 variables:
\describe{
  \item{\code{tractce}}{Current census tract code}
  \item{\code{geoid}}{Current nation-based census tract identifier; a concatenation of current state FIPS code, county FIPS code, and census tract number}
  \item{\code{name}}{Variable length geographic area name}
  \item{\code{namelsad}}{Current name and the translated legal/statistical area description code for census tract}
  \item{\code{aland}}{Current land area (square meters)}
  \item{\code{awater}}{Current water area (square meters)}
  \item{\code{intptlat}}{Current latitude of the internal point}
  \item{\code{intptlon}}{Current longitude of the internal point}
  \item{\code{geometry}}{Polygon with tract boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
baltimore_tracts
}
\description{
...
}
\keyword{datasets}
