% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_water}
\alias{baltimore_water}
\title{Baltimore Rivers and Streams}
\format{
A data frame with 947 rows and 2 variables:
\describe{
  \item{layer}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/maryland-waterbodies-rivers-and-streams-detailed}
}
\usage{
baltimore_water
}
\description{
Detailed line features of rivers and streams within a generalized boundary of Baltimore City, Maryland.
}
\keyword{datasets}
