% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_water}
\alias{baltimore_water}
\title{Baltimore Water}
\format{
A data frame with 468 rows and 6 variables:
\describe{
  \item{\code{name}}{Water feature name, if available}
  \item{\code{type}}{Water type}
  \item{\code{subtype}}{Water subtype}
  \item{\code{symbol}}{Symbol}
  \item{\code{water}}{Water indicator}
  \item{\code{geometry}}{Multipolygon geometry}
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/maryland-waterbodies-rivers-and-streams-detailed}
}
\usage{
baltimore_water
}
\description{
Detailed multipolygon data for streams, lakes, and other water in Baltimore City.
}
\keyword{datasets}
