% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_water}
\alias{baltimore_water}
\title{Baltimore Water}
\format{
A data frame with 468 rows and 6 variables:
\describe{
  \item{name}{Name if available}
  \item{type}{Water type}
  \item{subtype}{Water subtype}
  \item{symbol}{Symbol}
  \item{water}{Water indicator}
  \item{geometry}{Multipolygon geometry}
  ...
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/maryland-waterbodies-rivers-and-streams-detailed}
}
\usage{
baltimore_water
}
\description{
Detailed multipolygon data for streams, lakes, and other water in Baltimore City.
}
\keyword{datasets}
