% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_zones}
\alias{bcps_zones}
\title{Baltimore City Public Schools School Zones or School Attendance Zones}
\format{
A data frame with 96 rows and 4 variables:
\describe{
  \item{program_name}{Program or school name}
  \item{program_number}{Program number}
  \item{zone_name}{Program name with zone appended}
  \item{geometry}{Multipolygons with school zone boundaries}
  ...
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/ArcGIS/rest/services/BCPSZones_2021/FeatureServer/0}
}
\usage{
bcps_zones
}
\description{
Baltimore City Public Schools School Zones also known as School Attendance Zones.
}
\keyword{datasets}
