% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_zones}
\alias{bcps_zones}
\title{Baltimore City Public Schools School Zones or School Attendance Zones  (SY
2021-2022)}
\format{
A data frame with 96 rows and 4 variables:
\describe{
\item{\code{zone_name}}{Program name with zone appended}
\item{\code{program_number}}{Program number}
\item{\code{program_name_short}}{Program or school name (short)}
\item{\code{type}}{Program type}
\item{\code{category}}{Program category or grade band, e.g. E, EM, H, etc.}
\item{\code{geometry}}{Multipolygons with school zone boundaries}
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/ArcGIS/rest/services/SY2122_Ezones_and_Programs/FeatureServer/15}
}
\usage{
bcps_zones
}
\description{
Baltimore City Public Schools School Zones also known as School Attendance
Zones.
}
\keyword{datasets}
