% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{cache_baltimore_data}
\alias{cache_baltimore_data}
\title{Cache data for mapbaltimore package}
\usage{
cache_baltimore_data(data = NULL, filename = NULL, overwrite = FALSE)
}
\arguments{
\item{data}{Data to cache.}

\item{filename}{File name to use for cached file. Defaults to name of data.
If the data is an sf object make sure to include the file type, e.g.
"data.gpkg", supported by `sf::write_sf()`. All other data is written to
rda with `readr::write_rds()`.}

\item{overwrite}{Logical. Default FALSE. If TRUE, overwrite any existing cached files that use the same filename.}
}
\description{
Cache data to `rappdirs::user_cache_dir("mapbaltimore")`
}
