% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{cache_baltimore_data}
\alias{cache_baltimore_data}
\alias{cache_msa_streets}
\alias{cache_edge_of_pavement}
\alias{show_cached_files}
\title{Cache data for mapbaltimore package}
\usage{
cache_baltimore_data(data = NULL, filename = NULL, overwrite = FALSE)

cache_msa_streets(
 
    url = "https://geodata.md.gov/imap/rest/services/Transportation/MD_HighwayPerformanceMonitoringSystem/MapServer/2",
  filename = "baltimore_msa_streets.gpkg",
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  overwrite = FALSE
)

cache_edge_of_pavement(
 
    url = "https://gisdata.baltimorecity.gov/egis/rest/services/OpenBaltimore/Edge_of_Pavement/FeatureServer/0",
  filename = "edge_of_pavement.gpkg",
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  overwrite = FALSE
)

show_cached_files()
}
\arguments{
\item{data}{Data to cache.}

\item{filename}{File name to use for cached file. Defaults to name of data.
If the data is an sf object make sure to include the file type, e.g.
"data.gpkg", supported by \code{sf::write_sf()}. All other data is written to
rda with \code{readr::write_rds()}.}

\item{overwrite}{Logical. Default FALSE. If TRUE, overwrite any existing cached files that use the same filename.}
}
\value{
\code{show_cached_files()} returns a tibble with the columns:
\itemize{
\item \code{file}, the name of the file,
\item \code{size_MB}, file size in MB,
\item \code{modified}, date and time last modified
}
}
\description{
Cache data to \code{rappdirs::user_cache_dir("mapbaltimore")}
}
\details{
\itemize{
\item Use \code{cache_msa_streets()} to download and cache street centerline data for all counties in the Baltimore metropolitan area.
\item Use  \code{cache_edge_of_pavement()} to download and cache edge of pavement data for Baltimore city.
}
}
