% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_area}
\alias{check_area}
\title{Validate area provided to mapping or charting function.}
\usage{
check_area(area)
}
\arguments{
\item{area}{sf object. Must have a name column unless an \code{area_label} is provided.}
}
\description{
Validate an area for a mapping function or call the \code{get_area} function.
}
