% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_area.R
\name{check_area}
\alias{check_area}
\title{Validate area provided to mapping or charting function.}
\usage{
check_area(area)
}
\arguments{
\item{area}{\code{sf} object with a column named "name."}
}
\description{
Validate an area for a mapping function or another mapbaltimore function.
}
