% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_area.R
\name{clip_area}
\alias{clip_area}
\title{Clip an area to a portion of the whole area}
\usage{
clip_area(
  area,
  clip = c("top", "right", "bottom", "left", "topright", "bottomright", "bottomleft",
    "topleft"),
  flip = FALSE,
  edge = TRUE,
  edge_dist = 5
)
}
\arguments{
\item{area}{\code{sf} object to clip}

\item{clip}{Character string describing the part of the area to clip or
remove. Options include c("top", "right", "bottom", "left", "topright",
"bottomright", "bottomleft", "topleft"). If NULL, the area is not clipped
and a full edge can be returned.}

\item{flip}{Logical. Default FALSE. If TRUE, the reverse of the select area
is removed, e.g. if \code{clip} is "topright" the "bottomleft" area is removed
instead.}

\item{edge}{Logical. Default TRUE. If TRUE, only the edge of the clipped area
is returned.  If TRUE with a negative \code{edge_dist}, only the edges are kept
(center is removed). If TRUE with a positive \code{edge_dist}, the full area is
removed and a buffer kept. If FALSE, the full clipped area is returned.}

\item{edge_dist}{Numeric. Distance in meters to use for the edge. Default 5
meters. Use negative values for an inside edge or positive numbers for an
outside edge.}
}
\value{
\code{sf} object clipped based on parameters
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because the functionality has been incorporated
into the improved \code{\link[sfext:st_clip]{sfext::st_clip()}} function which uses a similar set
of parameters.

Clip based on the corner of the bounding box. Used for the street name
location option within \code{layer_area_streets}
}
\keyword{internal}
