% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_area.R
\name{clip_area}
\alias{clip_area}
\title{Clip an area to a portion of the whole area}
\usage{
clip_area(
  area,
  clip = c("edge", "top", "right", "bottom", "left", "topright", "bottomright",
    "bottomleft", "topleft"),
  flip = FALSE,
  edge_dist = 5
)
}
\arguments{
\item{area}{`sf` object to clip}

\item{clip}{Character string describing the part of the area to clip or
remove (except if clip is "edge"). If "edge" with a negative "edge_dist",
only the edges are kept (center is removed). If "edge" with a positive
"edge_dist", the full area is removed and a buffer kept. Options include
c("edge", "top", "right", "bottom", "left", "topright", "bottomright",
"bottomleft", "topleft"). Expect edge to move to a separate parameter
allowing clip and edge to be combined.}

\item{flip}{Logical. Default FALSE. If TRUE, the reverse of the select area
is removed, e.g. if clip is "topright" the "bottomleft" area is removed
instead.}

\item{edge_dist}{Numeric. Distance in meters to use for the edge. Default 5
meters. Use negative values for an inside edge or positive numbers for an
outside edge.}
}
\value{
\code{sf} object clipped based on parameters
}
\description{
Clip based on the corner of the bounding box. Used for the street name
location option within \code{layer_area_streets}
}
