% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congressional_districts}
\alias{congressional_districts}
\title{U.S. Congressional Districts for Baltimore City}
\format{
A data frame with 3 rows and 15 variables:
\describe{
  \item{\code{statefp}}{2-character state FIPS code}
  \item{\code{cd116fp}}{...}
  \item{\code{geoid}}{GeoID}
  \item{\code{namelsad}}{concatenated variable length geographic area name and legal/statistical area description (LSAD)}
  \item{\code{lsad}}{Legal/statistical area description (LSAD)}
  \item{\code{cdsessn}}{...}
  \item{\code{mtfcc}}{5-digit MAF/TIGER Feature Class Code (MTFCC)}
  \item{\code{funcstat}}{Current functional status}
  \item{\code{aland}}{Current land area (square meters)}
  \item{\code{awater}}{Current water area (square meters)}
  \item{\code{intptlat}}{Current latitude of the internal point}
  \item{\code{intptlon}}{Current longitude of the internal point}
  \item{\code{label}}{Congressional District label}
  \item{\code{name}}{Congressional District name}
  \item{\code{geometry}}{Multipolygon with Congressional district boundary}
}
}
\source{
\url{...}
}
\usage{
congressional_districts
}
\description{
U.S. Congressional Districts overlapping with Baltimore City. Downloaded with
the tigris package.
}
\keyword{datasets}
