% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congressional_districts}
\alias{congressional_districts}
\title{Baltimore City Congressional Districts}
\format{
A data frame with 3 rows and 15 variables:
\describe{
  \item{statefp}{...}
  \item{cd116fp}{...}
  \item{geoid}{...}
  \item{namelsad}{...}
  \item{lsad}{...}
  \item{cdsessn}{...}
  \item{mtfcc}{...}
  \item{funcstat}{...}
  \item{aland}{...}
  \item{awater}{...}
  \item{intptlat}{...}
  \item{intptlon}{...}
  \item{label}{...}
  \item{name}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{...}
}
\usage{
congressional_districts
}
\description{
U.S. Congressional Districts overlapping with Baltimore City. Downloaded with the tigris package.
}
\keyword{datasets}
