% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{council_districts}
\alias{council_districts}
\title{Baltimore City Council Districts}
\format{
A data frame with 14 rows and 2 variables:
\describe{
\item{\code{id}}{Number of the City Council district}
\item{\code{name}}{Name of the City Council district}
\item{\code{geometry}}{MULTIPOLYGON geometry with boundaries of City Council districts}
}
}
\source{
\url{https://geodata.baltimorecity.gov/egis/rest/services/CityView/City_Council_Districts/MapServer/0}
}
\usage{
council_districts
}
\description{
Boundaries for the Baltimore City Council Districts used since 2012
(following boundary revisions completed in 2011 based on the 2010 Decennial
Census).
}
\keyword{datasets}
