% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csas}
\alias{csas}
\title{Baltimore City Community Statistical Areas}
\format{
A data frame with 55 rows and 3 variables:
\describe{
  \item{id}{Community Statistical Area id number}
  \item{name}{Community Statistical Area name}
  \item{geometry}{Multipolygons with boundary of each area}
  ...
}
}
\source{
\url{...}
}
\usage{
csas
}
\description{
Baltimore City Community Statistical Areas (CSAs)
}
\keyword{datasets}
