% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csas}
\alias{csas}
\title{Community Statistical Areas (2010)}
\format{
A data frame with 55 rows and 3 variables:
\describe{
\item{\code{id}}{Community Statistical Area id number}
\item{\code{name}}{Community Statistical Area name}
\item{\code{url}}{URL to BNIA-JFI webpage on Community Statistical Area}
\item{\code{geometry}}{MULITPOLYGON boundary geometry}
}
}
\source{
\url{https://bniajfi.org/mapping-resources/}
}
\usage{
csas
}
\description{
Community Statistical Areas (CSAs) are clusters of neighborhoods and are
organized around U.S. Census tract boundaries by the Baltimore Neighborhood
Indicators Alliance. In some cases, CSA boundaries may cross neighborhood
boundaries. There are 55 CSAs in Baltimore City. Neighborhood lines often do
not fall along CSA boundaries. The CSAs were originally created in 2002 and
were revised for the publication of Vital Signs 10 using new 2010 Census
Tract boundaries. There are no anticipated boundary revisions in 2020.
}
\keyword{datasets}
