% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{explore_baltimore}
\alias{explore_baltimore}
\title{Explore Baltimore Heritage}
\format{
A data frame with 459 rows and 10 variables:
\describe{
  \item{id}{...}
  \item{featured}{...}
  \item{modified}{...}
  \item{title}{...}
  \item{address}{...}
  \item{thumbnail}{...}
  \item{fullsize}{...}
  \item{url}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{https://explore.baltimoreheritage.org/}
}
\usage{
explore_baltimore
}
\description{
A table of public stories on the Explore Baltimore Heritage website.
}
\keyword{datasets}
