% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{explore_baltimore}
\alias{explore_baltimore}
\title{Stories from Explore Baltimore Heritage}
\format{
A data frame with 459 rows and 10 variables:
\describe{
\item{\code{id}}{Story identifier}
\item{\code{featured}}{Featured indicator}
\item{\code{modified}}{Modified date/time}
\item{\code{title}}{Story title}
\item{\code{address}}{Street address for story location}
\item{\code{thumbnail}}{URL for thumbnail-size featured image}
\item{\code{fullsize}}{URL for full-size featured image}
\item{\code{url}}{URL for story}
\item{\code{geometry}}{Point for story location}
}
}
\source{
\url{https://explore.baltimoreheritage.org/}
}
\usage{
explore_baltimore
}
\description{
A table of public stories on the Explore Baltimore Heritage website.
}
\keyword{datasets}
