% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{explore_baltimore}
\alias{explore_baltimore}
\title{Explore Baltimore Heritage Stories}
\format{
A data frame with 491 rows and 10 variables:
\describe{
\item{\code{id}}{Story identifier}
\item{\code{featured}}{Featured indicator}
\item{\code{modified}}{Modified date/time}
\item{\code{title}}{Story title}
\item{\code{address}}{Street address for story location}
\item{\code{thumbnail}}{URL for thumbnail-size featured image}
\item{\code{fullsize}}{URL for full-size featured image}
\item{\code{url}}{URL for story}
\item{\code{geometry}}{POINT for story location}
}
}
\source{
\url{https://explore.baltimoreheritage.org/}
}
\usage{
explore_baltimore
}
\description{
A table of public stories on the \href{https://explore.baltimoreheritage.org/}{Explore Baltimore Heritage website} published by \href{https://baltimoreheritage.org/}{Baltimore Heritage}. The text of stories on Explore
Baltimore Heritage is licensed under a \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0 license}. Updated on 2023 March
29.
}
\keyword{datasets}
