% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_area}
\alias{get_area}
\title{Return geography for selected area type and name.}
\usage{
get_area(
  area_type = c("neighborhood", "council", "police", "csa", "blockgroup", "tract"),
  area_name = NULL,
  area_geoid = NULL,
  union = FALSE
)
}
\arguments{
\item{area_type}{Character vector of length 1. Required. Supported values include c("neighborhood", "council", "police", "csa", "blockgroup", "tract")}

\item{area_name}{Character vector of any length. Required if area_type is "neighborhood", "council", or "police"}

\item{area_geoid}{Character vector of any length. Required if area_type is "blockgroup" or "tract"}

\item{union}{Defaults to FALSE. If TRUE and multiple areas are provided, the area geography will be combined into a single sf object.}
}
\description{
Get the geometry and name of a selected neighborhood, City Council district, police district,
Community Statistical Area, U.S. Census Block group or U.S. Census tract.
}
