% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_area}
\alias{get_area}
\title{Return geography for selected area type and name.}
\usage{
get_area(
  type = c("neighborhood", "council district", "legislative district",
    "congressional district", "planning district", "police district", "csa"),
  area_name = NULL,
  area_id = NULL,
  union = FALSE
)
}
\arguments{
\item{type}{area type matching one of the included boundary datasets.
Supported values include c("neighborhood", "council district", "legislative district",
"congressional district", "planning district", "police district", "csa")}

\item{area_name}{name or names matching id column in data of selected dataset.}

\item{area_id}{identifier or identifiers matching id column of selected dataset.
Not all supported datasets have an id column}

\item{union}{If TRUE and multiple area names are provided, the area geometry is combined
with \code{\link[sf]{st_union}} and names are concatenated into a single string.
Defaults to FALSE.}
}
\description{
Get a sf object with one or more neighborhoods, Baltimore City Council districts,
Maryland Legislative Districts, U.S. Congressional Districts, Baltimore Planning Districts,
Baltimore Police Districts, or Community Statistical Areas.
}
