% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_area}
\alias{get_area}
\title{Return geography for selected area type and name.}
\usage{
get_area(
  area_type = c("neighborhood", "council_district", "police_district", "csa",
    "block_group", "tract"),
  area_name = NULL,
  union = FALSE
)
}
\arguments{
\item{area_type}{Character vector of length 1. Required. Supported values include c("neighborhood", "council", "police", "csa", "blockgroup", "tract")}

\item{area_name}{Character vector of any length. For U.S. Census geographies ("blockgroup", "tract"), provide selected geoid value or values instead of a name.}

\item{union}{Defaults to FALSE. If TRUE and multiple area names are provided, the area geometry is combined with \code{sf::st_union} and the area names is coerced from a vector to a nested list.}
}
\description{
Get the geometry and name of a selected neighborhood, City Council district, police district,
Community Statistical Area, U.S. Census Block group or U.S. Census tract.
}
