% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_911_calls.R
\name{get_area_911_calls}
\alias{get_area_911_calls}
\title{Get area 911 calls for service from Open Baltimore}
\usage{
get_area_911_calls(
  area_type = NULL,
  area_name = NULL,
  description = NULL,
  year = NULL,
  start_date = NULL,
  end_date = NULL,
  where = NULL,
  ...
)
}
\arguments{
\item{area_type}{Area type. Requires area_name is also provided. Options
include "neighborhood", "council district", or "police district"}

\item{area_name}{Area name. Requires area_type is also provided.}

\item{description}{String matching call description, e.g. "DISORDERLY",
"BURGLARY", "DISCHRG FIREARM", etc.}

\item{year}{numeric. Year of calls for service. Currently only one year at a
time is supported (except for 2021 and 2022). If NULL, the oldest year from
the start_date and end_date is used.}

\item{start_date}{Character string in format YYYY-MM-DD. Filters calls by
date.}

\item{end_date}{Character string in format YYYY-MM-DD.  Filters calls by
date.}

\item{where}{string for where condition. Ignored if area_type, area_name,
start_date, or end_date are provided.}

\item{...}{Additional parameters passed to \code{\link[getdata:get_esri_data]{getdata::get_esri_data()}}
excluding url, where, crs, and .name_repair.}
}
\description{
Get 911 calls for service from 2017 through the present year.
}
