% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_bcps_programs.R
\name{get_area_bcps_programs}
\alias{get_area_bcps_programs}
\title{Get BCPS programs and attendance zones for a local area}
\usage{
get_area_bcps_programs(area)
}
\arguments{
\item{area}{Name of the neighborhood in sentence case.}
}
\description{
Get BCPS programs and attendance zones for a local area
}
\details{
Returns a named list with overlapping BCPS attendance zones, program
locations associated with those zones, and any additional programs located
within the area.
}
