% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_citations.R
\name{get_area_citations}
\alias{get_area_citations}
\title{Get area citations}
\usage{
get_area_citations(
  area_type = NULL,
  area_name = NULL,
  description = NULL,
  crs = 2804
)
}
\arguments{
\item{area_type}{Area type. Options include "neighborhood", "council district", or "police district"}

\item{area_name}{Area name.}

\item{description}{String to match to citation description, e.g. "SIGNS" matches "PROHIBITED POSTING OF SIGNS ON PUBLIC PROPERTY"}

\item{crs}{Coordinate reference system. Default 2804.}
}
\description{
Get Environmental Control Board (ECB) citations.
}
