% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_crime.R
\name{get_area_crime}
\alias{get_area_crime}
\title{Get area crimes from Open Baltimore}
\usage{
get_area_crime(
  area,
  description = NULL,
  where = "1=1",
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804)
)
}
\arguments{
\item{area}{\code{sf} object. Optional. Only used if trim is TRUE.}

\item{description}{Crime type or description. Supported options include "AGG.
ASSAULT", "ARSON", "AUTO THEFT", "BURGLARY", "COMMON ASSAULT", "HOMICIDE",
"LARCENY", "LARCENY FROM AUTO", "RAPE", "ROBBERY - CARJACKING", "ROBBERY -
COMMERCIAL", "ROBBERY - RESIDENCE", "ROBBERY - STREET", or "SHOOTING"}

\item{where}{string for where condition. This parameter is ignored if a
description is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{Logical. Default \code{FALSE}. If \code{TRUE}, area is required.}

\item{crs}{Coordinate reference system. Default 2804.}
}
\description{
Get reported crimes since 2014 for a specific area.
}
\examples{
\dontrun{
# Get shootings for the Lauraville area
area <- get_area("neighborhood", "Lauraville")
crimes <-
 get_area_crime(area = area, description = "SHOOTING")
}
}
