% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_data.R
\name{get_area_data}
\alias{get_area_data}
\title{Get local or cached data for an area}
\usage{
get_area_data(
  area = NULL,
  bbox = NULL,
  data = NULL,
  extdata = NULL,
  cachedata = NULL,
  path = NULL,
  url = NULL,
  fn = NULL,
  diag_ratio = NULL,
  dist = NULL,
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  crs = NULL
)
}
\arguments{
\item{area}{\code{sf} object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf]{st_union}}}

\item{bbox}{\code{bbox} object defining area used to filter data. If an area is
provided, the bounding box is ignored.}

\item{data}{\code{sf} object including data in area}

\item{extdata}{Character. Name of an external geopackage (.gpkg) file
included with the package where selected data is available. Available data
includes "trees", "unimproved_property", and "vegetated_area"}

\item{cachedata}{Character. Name of a cached geopackage (.gpkg) file where
selected data is available. Running \code{cache_mapbaltimore_data()} caches
data for "real_property", "baltimore_msa_streets", and "edge_of_pavement"}

\item{path}{Character. Path to local or remote spatial data file supported by
\code{\link[sf]{st_read}}}

\item{url}{Character. URL for FeatureServer or MapServer layer to pass to get_area_esri_data.}

\item{fn}{Function to apply to area data before returning.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crop}{If TRUE, data cropped to area or bounding box
\code{\link[sf]{st_crop}} adjusted by the `dist`, `diag_ratio`, and `asp`
parameters provided. Default \code{TRUE}.}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf]{st_intersection}}. This option is not supported for any
adjusted areas that use the `dist`, `diag_ratio`, or `asp` parameters.
Default \code{FALSE}.}

\item{crs}{Coordinate Reference System (CRS) to use for the returned data.
The CRS of the provided data and bounding box or area must match one
another but are not required to match the CRS provided by this parameter.}
}
\description{
Returns data for a selected area or areas with an optional buffer. If both
crop and trim are FALSE, the function uses \code{\link[sf]{st_intersects}} to
filter data to include the full geometry of anything that overlaps with the
area or bbox (if the area is not provided).
}
