% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_esri_data.R
\name{get_area_esri_data}
\alias{get_area_esri_data}
\title{Get data from an ArcGIS FeatureServer or MapServer}
\usage{
get_area_esri_data(
  area = NULL,
  bbox = NULL,
  url = NULL,
  where = "1=1",
  type = c("md food stores 2017 2018", "farmers markets 2020",
    "baltimore food stores 2016", "baltimore demolitions", "contour 2ft",
    "contours 10ft", "open vacant building notices", "liquor licenses",
    "fixed speed cameras", "red light cameras", "edge of pavement"),
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804)
)
}
\arguments{
\item{area}{\code{sf} object. Optional. Only used if trim is TRUE.}

\item{bbox}{\code{bbox} object. Optional but suggested to avoid downloading
entire layer. See \code{\link[sf]{st_bbox}} for more information.}

\item{url}{FeatureServer or MapServer url to retrieve data from. Passed to
\code{url} parameter of \code{esri2sf::esri2sf()} function.}

\item{where}{string for where condition. Default is 1=1 for all rows.}

\item{type}{Type of data to get. Options include "md food stores 2017 2018",
"farmers markets 2020", "baltimore food stores 2016", "baltimore
demolitions", "contour 2ft", "contours 10ft", "open vacant building
notices", "liquor licenses", "fixed speed cameras", "red light cameras",
and "edge of pavement"}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{Logical. Default \code{FALSE}. If \code{TRUE}, area is required.}

\item{crs}{Coordinate reference system. Default 2804.}
}
\description{
Wraps the \code{esri2sf::esri2sf()} function to download an ArcGIS
FeatureServer or MapServer.
}
