% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_osm_data.R
\name{get_area_osm_data}
\alias{get_area_osm_data}
\title{Get Open Street Map features for area}
\usage{
get_area_osm_data(
  area = NULL,
  bbox = NULL,
  key,
  value = NULL,
  return_type = c("osm_polygons", "osm_points", "osm_lines", "osm_multilines",
    "osm_multipolygons"),
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804)
)
}
\arguments{
\item{area}{sf object. If multiple areas are provided, they are unioned into
a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}}

\item{bbox}{\code{bbox} object to buffer and/or adjust. If an area is provided, any bbox is ignored.}

\item{key}{feature key for overpass query. If key is "building" and value is
NULL, a preset list of tag values will be used to return all available
buildings in the bounding box.}

\item{value}{for feature key; can be negated with an initial exclamation
mark, value = "!this", and can also be a vector, value = c ("this",
"that").}

\item{return_type}{Character vector length 1 with geometry type to return.
Defaults to returningpolygons. Set to NULL to return all types.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crop}{Logical. Default TRUE. If TRUE, use the \code{\link[sf:st_crop]{sf::st_crop()}}
to trim results to area bounding box.}

\item{trim}{Logical. Default FALSE. If TRUE, use the
\code{\link[sf:geos_binary_ops]{sf::st_intersection()}} function to trim results to area polygon.}

\item{crs}{EPSG code for the coordinate reference system for the plot.
Default is 2804. See \url{https://epsg.io/} for more information.}
}
\description{
Wraps \code{osmdata} functions.
}
