% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_permits.R
\name{get_area_permits}
\alias{get_area_permits}
\title{Get area permits from Open Baltimore}
\usage{
get_area_permits(
  area,
  year = 2021,
  permit_type = NULL,
  where = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804)
)
}
\arguments{
\item{area}{\code{sf} object. Optional. Only used if trim is TRUE.}

\item{year}{Year. Must be 2019 or later.}

\item{permit_type}{Optional. Supported values include "USE", "DEM", "COM", or
"BMZ".}

\item{where}{string for where condition. permit_type and year are ignored if
a custom `where` is provided. Set where to "1=1" to return data for all
years since 2019.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{Logical. Default \code{FALSE}. If \code{TRUE}, area is required.}

\item{crs}{Coordinate reference system. Default 2804.}
}
\description{
Get building permits from 2019 through the present.
}
