% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_permits.R
\name{get_area_permits}
\alias{get_area_permits}
\title{Get area building permits from Open Baltimore}
\usage{
get_area_permits(
  area,
  year = 2022,
  date_range = NULL,
  permit_type = NULL,
  where = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = "m",
  asp = NULL,
  trim = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  ...
)
}
\arguments{
\item{year}{Year. Must be 2019 or later.}

\item{date_range}{Date range as character vector in format of c("YYYY-MM-DD",
"YYYY-MM-DD"). Minimum and maximum values are used if length is greater
than 1.}

\item{permit_type}{Optional. Supported values include "USE", "DEM", "COM", or
"BMZ".}

\item{where}{string for where condition. permit_type and year are ignored if
a custom \code{where} is provided. Set where to "1=1" to return data for all
years since 2019.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{trim}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}

\item{crs}{Cordinate reference system to return, Default: 4326 for
\code{\link[sfext:sf_to_df]{sf_to_df()}} and \code{NULL} for \code{\link[sfext:df_to_sf]{df_to_sf()}}.}

\item{...}{Additional parameters passed to \code{\link[getdata:get_esri_data]{getdata::get_esri_data()}}.}
}
\description{
Get building permits from 2019 through the present.
}
