% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_property.R
\name{get_area_property}
\alias{get_area_property}
\alias{format_property_data}
\title{Get real property data}
\usage{
get_area_property(
  area = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  cache = FALSE,
  ...
)

format_property_data(data)
}
\arguments{
\item{area}{\code{sf} object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}}

\item{bbox}{\code{bbox} object defining area used to filter data. If an area is
provided, the bounding box is ignored.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crop}{If TRUE, data cropped to area or bounding box
\code{\link[sf:st_crop]{sf::st_crop()}} adjusted by the \code{dist}, \code{diag_ratio}, and \code{asp}
parameters provided. Default \code{TRUE}.}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf:geos_binary_ops]{sf::st_intersection()}}. This option is not supported for any
adjusted areas that use the \code{dist}, \code{diag_ratio}, or \code{asp} parameters.
Default \code{FALSE}.}

\item{cache}{If TRUE, cache data to mapbaltimore cache folder. Default FALSE.}

\item{...}{Use to pass filename and overwrite parameter to cache_baltimore_data. Use gpkg file type.}

\item{data}{\code{sf} object including data in area}
}
\description{
Get showing parcels described as owner occupied, non-owner occupied, vacant,
and unimproved. Real property or parcel data is from the Maryland State
Department of Assessment and Taxation and may include errors.
}
