% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_property.R
\name{get_area_property}
\alias{get_area_property}
\title{Get real property data for an area}
\usage{
get_area_property(
  area,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE
)
}
\arguments{
\item{area}{Required sf class tibble. Must include a name column.}

\item{bbox}{bbox object to buffer and/or adjust. If an area is provided, any bbox is ignored.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").}
}
\description{
Map showing parcels described as owner occupied, non-owner occupied, vacant, and unimproved.
Real property or parcel data is from the Maryland State Department of Assessment and Taxation and may include errors. Check the \code{mapbaltimore::real_property} data description for details.
}
