% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_property.R
\name{get_area_property}
\alias{get_area_property}
\title{Get real property data for an area}
\usage{
get_area_property(
  area,
  dist = NULL,
  diag_ratio = NULL,
  trim = FALSE,
  area_name = NULL,
  type = c("block", "neighborhood", "police district", "council district", "csa",
    "block group", "tract")
)
}
\arguments{
\item{area}{Required sf class tibble. Must include a name column.}

\item{area_name}{Name of the selected type to provide. If}

\item{type}{Optional character vector for the type of area (e.g. "neighborhood", "block_group"). The name column for the provided \code{area} must match the names or geoid of the one or more areas of the provided type to return data.
If buffer is TRUE, the type is ignored and the \code{\link[sf]{st_crop}} function is used.}

\item{buffer}{Optional. If default (NULL), the returned real property data includes property within a default buffered distance (1/8th of the diagonal distance across the bounding box). If numeric, the function returns data cropped to area buffered by this distance in meters.}
}
\description{
Map showing parcels described as owner occupied, non-owner occupied, vacant, and unimproved.
Real property or parcel data is from the Maryland State Department of Assessment and Taxation and may include errors. Check the \code{mapbaltimore::real_property} data description for details.
}
