% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_requests.R
\name{get_area_requests}
\alias{get_area_requests}
\title{Get area 311 service requests from Open Baltimore}
\usage{
get_area_requests(
  area,
  year = 2022,
  date_range = NULL,
  request_type = NULL,
  agency = NULL,
  where = "1=1",
  dist = NULL,
  diag_ratio = NULL,
  unit = "m",
  asp = NULL,
  trim = FALSE,
  geometry = TRUE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  duplicates = FALSE
)
}
\arguments{
\item{year}{Year for service requests. Default 2021. 2017 to 2022 supported.}

\item{request_type}{Service request type.}

\item{agency}{City agency responsible for request. Options include
"Transportation", "BGE", "Solid Waste", "Housing", "Water Wastewater",
"Health", "Call Center", "Finance", "Liquor Board", "Recreation & Parks",
"Fire Department", "Parking Authority", and "General Services"}

\item{where}{string for where condition. This parameter is ignored if a
request_type or agency are provided.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{geometry}{Default \code{TRUE.} If \code{FALSE}, return requests with missing
latitude/longitude (for years prior to 2021 only).}

\item{crs}{Cordinate reference system to return, Default: 4326 for
\code{\link[sfext:sf_to_df]{sf_to_df()}} and \code{NULL} for \code{\link[sfext:df_to_sf]{df_to_sf()}}.}

\item{duplicates}{If \code{TRUE}, return 311 service requests marked as
"Duplicate". If \code{FALSE}, filter duplicate requests out of results.}
}
\description{
Get 311 service requests for a specific area. Service requests
from 2017 to 2020 area available but only a single year can be requested at
a time. Duplicate requests are removed from the returned data. Requests can
be filtered by request type, responsible city agency, or both. You can
return multiple types or agencies, by using a custom where query parameter
or by calling each type/agency separately.
}
\examples{
# Get boundary for Edmondson Village
area <- get_area("neighborhood", "Edmondson Village")

# Get fallen limb requests for 2022
get_area_requests(
  area = area,
  date_range = c("2022-11-01", "2022-12-31"),
  request_type = "FOR-Fallen Limb"
)

# Get dirty alley service requests for multiple years using purrr::map_dfr()
purrr::map_dfr(
  c(2021, 2020),
  ~ get_area_requests(
    area = area,
    year = .x,
    request_type = "SW-Dirty Alley"
  )
)
}
