% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_requests.R
\name{get_area_requests}
\alias{get_area_requests}
\title{Get area 311 service requests}
\usage{
get_area_requests(
  area,
  year = 2021,
  request_type = NULL,
  agency = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  crs = 2804
)
}
\arguments{
\item{area}{\code{sf} object. Optional. Only used if trim is TRUE.}

\item{year}{Year for service requests (2021 is currently the only year supported), Default: 2021}

\item{request_type}{Service request type.}

\item{agency}{City agency assigned for request. Options include "Transportation", "BGE", "Solid Waste", "Housing", "Water Wastewater", "Health", "Call Center", "Finance", "Liquor Board", "Recreation & Parks", "Fire Department", "Parking Authority", and "General Services"}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{Logical. Default \code{FALSE}. If \code{TRUE}, area is required.}

\item{crs}{Coordinate reference system. Default 2804.}
}
\description{
Get 311 service requests for a specific area.
}
