% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_requests.R
\name{get_area_requests}
\alias{get_area_requests}
\title{Get area 311 service requests from Open Baltimore}
\usage{
get_area_requests(
  area,
  year = 2021,
  request_type = NULL,
  agency = NULL,
  where = "1=1",
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  geometry = TRUE,
  crs = 2804
)
}
\arguments{
\item{area}{\code{sf} object. Optional. Only used if trim is TRUE.}

\item{year}{Year for service requests. Default 2021. 2017 to 2021 supported.}

\item{request_type}{Service request type.}

\item{agency}{City agency responsible for request. Options include
"Transportation", "BGE", "Solid Waste", "Housing", "Water Wastewater",
"Health", "Call Center", "Finance", "Liquor Board", "Recreation & Parks",
"Fire Department", "Parking Authority", and "General Services"}

\item{where}{string for where condition. This parameter is ignored if a
request_type or agency are provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{Logical. Default \code{FALSE}. If \code{TRUE}, area is required.}

\item{geometry}{Default TRUE. If FALSE, return requests with missing
latitude/longitude for years prior to 2021.}

\item{crs}{Coordinate reference system. Default 2804.}
}
\description{
Get 311 service requests for a specific area. Service requests
  from 2017 to 2020 area available but only a single year can be requested at
  a time. Duplicate requests are removed from the returned data. Requests can
  be filtered by request type, responsible city agency, or both. You can
  return multiple types or agencies, by using a custom where query parameter
  or by calling each type/agency separately.
}
\examples{
\dontrun{
# Get dirty alley service requests for Edmondson Village in 2020 and 2021
area <- get_area("neighborhood", "Edmondson Village")
requests <-
  map_dfr(
    c(2021, 2020),
    ~ get_area_requests(area = area, year = .x, request_type = "SW-Dirty Alley")
  )
}
}
