% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_streetnames.R
\name{get_area_streets}
\alias{get_area_streets}
\title{Get area streets}
\usage{
get_area_streets(area, street_type = NULL, sha_class = NULL)
}
\arguments{
\item{area}{sf class tibble. Object must include a name column.}

\item{street_type}{Character vector with list of street subtypes to include. By default, includes all subtypes except alleys ("STRALY").}

\item{sha_class}{Character vector. "all" selects all streets with an assigned SHA classification. Options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")}
}
\description{
Get streets within an area or areas.
}
