% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streets.R
\name{get_area_streets}
\alias{get_area_streets}
\title{Get area streets}
\usage{
get_area_streets(
  area,
  street_type = NULL,
  sha_class = NULL,
  dist = NULL,
  diag_ratio = NULL,
  trim = FALSE,
  msa = FALSE
)
}
\arguments{
\item{area}{sf object with area of streets to return.}

\item{street_type}{selected street subtypes to include. Includes all subtypes except alleys ("STRALY") by default.
Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN")}

\item{sha_class}{selected SHA classifications to include.
"all" selects all streets with an assigned SHA classification (around one-quarter of all street segments).
Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{trim}{Logical. Default FALSE. Trim streets to area using sf::st_intersection().}

\item{msa}{Logical. Default FALSE. Get streets from cached baltimore_msa_streets.gpkg file.}
}
\description{
Get streets within an area or areas.
}
