% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_streets.R
\name{get_area_streets}
\alias{get_area_streets}
\title{Get selected area streets}
\usage{
get_area_streets(
  area = NULL,
  street_type = NULL,
  sha_class = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  msa = FALSE,
  union = TRUE
)
}
\arguments{
\item{area}{sf object with area of streets to return.}

\item{street_type}{selected street subtypes to include. By default, the returned data includes all subtypes except alleys ("STRALY").
Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN")}

\item{sha_class}{selected SHA classifications to include.
"all" selects all streets with an assigned SHA classification (around one-quarter of all street segments).
Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")}

\item{bbox}{\code{bbox} object defining area used to filter data. If an area is
provided, the bounding box is ignored.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{Logical. Default \code{FALSE}. Trim streets to area using `sf::st_intersection()`.}

\item{msa}{Logical. Default \code{FALSE}. Get streets from cached `baltimore_msa_streets.gpkg` file using `cachedata` parameter of `get_area_data` function.}

\item{union}{Logical. Default \code{TRUE}. Union geometry based on `fullname` of streets.}
}
\description{
Get streets within an area or areas.
}
