% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning.R
\name{get_area_zoning}
\alias{get_area_zoning}
\title{Get zoning data a local area}
\usage{
get_area_zoning(
  area,
  category = c("all", "residential", "commercial", "industrial"),
  diag_ratio = 0.125,
  dist = NULL,
  trim = FALSE,
  crs = 2804
)
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{category}{Zoning category to return. "all", "residential", "commercial", "industrial"}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{trim}{If TRUE, data trimmed to area with \code{\link[sf]{st_intersection()}}. Default FALSE.}

\item{crs}{Selected CRS for returned data}
}
\description{
Get zoning codes for an area within the city.
The 2017 zoning data does not include any exemptions granted by the BMZA (Board of Municipal Zoning Appeals).
}
