% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_zoning.R
\name{get_area_zoning}
\alias{get_area_zoning}
\title{Get zoning data for an area}
\usage{
get_area_zoning(
  area = NULL,
  bbox = NULL,
  category = c("all", "residential", "commercial", "industrial"),
  diag_ratio = NULL,
  dist = NULL,
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  crs = NULL,
  union = FALSE
)
}
\arguments{
\item{area}{Required \code{sf} object with a 'name' column.}

\item{bbox}{\code{bbox} object defining area used to filter data. If an area is
provided, the bounding box is ignored.}

\item{category}{Zoning category to return. "all", "residential", "commercial", "industrial"}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crop}{If TRUE, data cropped to area or bounding box
\code{\link[sf]{st_crop}} adjusted by the `dist`, `diag_ratio`, and `asp`
provided. Default \code{TRUE}.}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf]{st_intersection}}. This option is not supported for any
adjusted areas that use the `dist`, `diag_ratio`, or `asp` parameters.
Default \code{FALSE}.}

\item{crs}{Coordinate Reference System (CRS) to use for the returned data.
The CRS of the provided data and bounding box or area must match one
another but are not required to match the CRS provided by this parameter.}

\item{union}{Logical. Default FALSE. If true, group zoning by label and combine geometry with \code{\link[sf]{st_union}}.}
}
\value{
\code{sf} object with zoning and overlay data for area.
}
\description{
Get zoning codes for an area within a provided sf or bbox object.
}
\details{
This 2017 zoning data does not include any exemptions granted by the Baltimore
City BMZA (Board of Municipal Zoning Appeals).
}
