% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baltimore_worker_flows.R
\name{get_baltimore_worker_flows}
\alias{get_baltimore_worker_flows}
\title{Get Baltimore metro area worker flows from the Census Transportation Planning
data (2012-2016 ACS)}
\usage{
get_baltimore_worker_flows(
  area,
  tracts = baltimore_tracts,
  min_estimate = 10,
  geometry = TRUE,
  crs = 2804
)
}
\arguments{
\item{area}{A sf or sfc object that intersects with tracts.}

\item{tracts}{Data from \code{\link[tigris:tracts]{tigris::tracts()}} for one or more county in the
Balitmore metro area. Defaults to \code{baltimore_tracts}.}

\item{min_estimate}{Minimum number of workers or residents a tract must have
to include in results. Tracts with fewer than the min_estimate values are
filtered out of results. Defaults to 10.}

\item{geometry}{If \code{TRUE}, return a list of sf objects. If \code{FALSE}, return a
list of data.frame objects. Defaults to \code{TRUE}.}

\item{crs}{Coordinate reference system to use for returned data when
\code{geometry = TRUE}. Defaults to 2804.}
}
\value{
A list of two data.frames or sf objects named "to" and "from".
}
\description{
Use FeatureLayers provided by the Baltimore Metropolitan Council.
}
