% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_buffered_area}
\alias{get_buffered_area}
\title{Get buffered area}
\usage{
get_buffered_area(area, buffer_distance = NULL)
}
\arguments{
\item{area}{sf object.}

\item{buffer_distance}{Optional. A single numeric vector representing the buffer distance in meters.}
}
\description{
Return an sf object of an area with a buffer applied to it. If no buffer distance is provided, a default buffer is calculated of one-eighth the diagonal distance of the bounding box (corner to corner) for the area. The metadata for the provided area remains the same.
}
