% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maryland_open_data.R
\name{get_maryland_open_resource}
\alias{get_maryland_open_resource}
\title{Get dataset from Maryland Open Data portal with optional SoQL parameters}
\usage{
get_maryland_open_resource(
  resource = NULL,
  select = NULL,
  where = NULL,
  query = NULL,
  geometry = FALSE
)
}
\arguments{
\item{resource}{Character vector with the Socrata dataset identifier for the dataset on Maryland's Open Data portal}

\item{select}{Character vector with SODA $select parameter. This is the set of columns to be returned, similar to a SELECT in SQL. \url{https://dev.socrata.com/docs/queries/select.html}}

\item{where}{Character vector with SODA $where parameter. Filters the rows to be returned, similar to WHERE. \url{https://dev.socrata.com/docs/queries/where.html}}

\item{query}{Character vector with SODA $query parameter. A full SoQL query string, all as one parameter. \url{https://dev.socrata.com/docs/queries/query.html}}

\item{geometry}{If this is set to TRUE and dataset contains a 'longitude' and 'latitude' column, the function returns a \code{sf} object.}
}
\description{
Get a selected Maryland Open dataset filtered by selected Socrata Query Language (SoQL) parameters as a tibble or sf object. Details on SoQL queries are found in the Socrata API documentation \url{https://dev.socrata.com/docs/queries/}
}
\examples{

\dontrun{
## Get Q2 2020 vehicle crash data for Cecil County, Maryland
get_maryland_open_resource(
  resource = "65du-s3qu",
  where = "(year = '2020') AND (quarter = 'Q2') AND county_desc like 'Cecil'"
)
}
}
