% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearby_areas.R
\name{get_nearby_areas}
\alias{get_nearby_areas}
\title{Get nearby areas}
\usage{
get_nearby_areas(
  area,
  type = c("neighborhood", "council district", "legislative district",
    "congressional district", "planning district", "police district", "csa"),
  dist = 1
)
}
\arguments{
\item{area}{sf object. Must have a name column unless an \code{area_label} is provided.}

\item{type}{Length 1 character vector. Required to match one of the supported area types (excluding U.S. Census types). This is the area type for the areas to return and is not required to be the same type as the provided area.}

\item{dist}{Distance in meters for matching nearby areas. Default is 1 meter.}
}
\description{
Return data for all areas of a specified type within a specified distance of another area.
}
