% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoning.R
\name{get_zoning}
\alias{get_zoning}
\title{Get zoning data a local area}
\usage{
get_zoning(area, buffer = FALSE, buffer_distance = 0)
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{buffer}{Logical. Defaults to FALSE. Option to get zoning data for the area and a surrounding buffer (TRUE) or only the area (FALSE).}

\item{buffer_distance}{Numeric. If no value is provided, the buffer is set as 1/8 (0.125) the diagonal corner-to-corner distance across the bounding box.}
}
\description{
Get zoning codes for an area within the city.
The 2017 zoning data does not include any exemptions grants by the BMZA (Board of Municipal Zoning Appeals).
}
