% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inspire_plans}
\alias{inspire_plans}
\title{INSPIRE Plans}
\format{
A data frame with 24 rows and 19 variables:
\describe{
\item{\code{plan_name}}{character COLUMN_DESCRIPTION}
\item{\code{plan_name_short}}{character COLUMN_DESCRIPTION}
\item{\code{overall_status}}{character COLUMN_DESCRIPTION}
\item{\code{recommendation_report_status}}{character COLUMN_DESCRIPTION}
\item{\code{adoption_status}}{character COLUMN_DESCRIPTION}
\item{\code{implementation_status}}{character COLUMN_DESCRIPTION}
\item{\code{inspire_lead_planner}}{character COLUMN_DESCRIPTION}
\item{\code{plan_url}}{character COLUMN_DESCRIPTION}
\item{\code{recommendation_report_url}}{character COLUMN_DESCRIPTION}
\item{\code{year_launched}}{character COLUMN_DESCRIPTION}
\item{\code{launch_date_target}}{character COLUMN_DESCRIPTION}
\item{\code{kick_off_presentation_date}}{character COLUMN_DESCRIPTION}
\item{\code{walking_route_id_target_date}}{character COLUMN_DESCRIPTION}
\item{\code{recommendations_date_target}}{character COLUMN_DESCRIPTION}
\item{\code{year_adopted}}{character COLUMN_DESCRIPTION}
\item{\code{commission_review_date_target}}{character COLUMN_DESCRIPTION}
\item{\code{adoption_date}}{character COLUMN_DESCRIPTION}
\item{\code{document_url}}{character COLUMN_DESCRIPTION}
\item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\usage{
inspire_plans
}
\description{
Data frame and boundary geometry for INSPIRE Plans adopted and in progress.
}
\details{
Note: this documentation is a placeholder.

DETAILS
}
\keyword{datasets}
