% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_area_data.R
\name{layer_area_data}
\alias{layer_area_data}
\title{Layer for area data}
\usage{
layer_area_data(
  area = NULL,
  bbox = NULL,
  data = NULL,
  extdata = NULL,
  cachedata = NULL,
  path = NULL,
  url = NULL,
  fn = NULL,
  asis = FALSE,
  diag_ratio = NULL,
  dist = NULL,
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  mapping = aes(),
  inherit.aes = TRUE,
  show_mask = FALSE,
  show_area = FALSE,
  area_aes = list(color = "gray30"),
  layer_after = NULL,
  ...
)
}
\arguments{
\item{area}{\code{sf} object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}}

\item{bbox}{\code{bbox} object defining area used to filter data. If an area is
provided, the bounding box is ignored.}

\item{data}{\code{sf} object including data in area}

\item{extdata}{Character. Name of an external geopackage (.gpkg) file
included with the package where selected data is available. Available data
includes "trees", "unimproved_property", and "vegetated_area"}

\item{cachedata}{Character. Name of a cached geopackage (.gpkg) file where
selected data is available. Running \code{cache_mapbaltimore_data()} caches
data for "real_property", "baltimore_msa_streets", and "edge_of_pavement"}

\item{path}{Character. Path to local or remote spatial data file supported by
\code{\link[sf:st_read]{sf::st_read()}}}

\item{url}{Character. URL for FeatureServer or MapServer layer to pass to get_area_esri_data.}

\item{fn}{Function to apply to area data before returning.}

\item{asis}{Logical. Default FALSE. If TRUE, use inherited data as is without
cropping to area.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crop}{If TRUE, data cropped to area or bounding box
\code{\link[sf:st_crop]{sf::st_crop()}} adjusted by the \code{dist}, \code{diag_ratio}, and \code{asp}
parameters provided. Default \code{TRUE}.}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf:geos_binary_ops]{sf::st_intersection()}}. This option is not supported for any
adjusted areas that use the \code{dist}, \code{diag_ratio}, or \code{asp} parameters.
Default \code{FALSE}.}

\item{crs}{Coordinate Reference System (CRS) to use for the returned data.
The CRS of the provided data and bounding box or area must match one
another but are not required to match the CRS provided by this parameter.}

\item{show_mask}{Logical. Default FALSE. If TRUE, add a mask using
\code{layer_area_mask}}

\item{show_area}{Logical. Default FALSE. If TRUE, add an outline of the area
to the layer.}

\item{area_aes}{List of fixed aesthetics for area layer. Default to
list(color = "gray30", fill = NA). Supported aesthetics include color,
fill, linetype, alpha, and size. Defaults to color = "gray30", fill = NA,
size = 0.75, alpha = 1, linetype = 0. Default aesthetics for geom_sf are
ignored.}

\item{layer_after}{ggplot objects to append to the end of the data, mask,
area layer list. This is intended to be used to append scales, guides, or
labels related to the the data layer.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsf]{ggplot2::geom_sf}}, \code{\link[ggplot2:ggsf]{ggplot2::geom_sf}}
  \describe{
    \item{\code{mapping}}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}
    \item{\code{inherit.aes}}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
  }}
}
\description{
Layer for ggplot to show data for an area or area bbox.
}
\details{
Combines get_area_data and geom_sf into a single call. Inherits data from
ggplot() if data, extdata, and cachedata are left as NULL. Set asis to TRUE
to keep data as is and not crop to area or modified area. Optionally can
combine data layer with a mask layer, an area layer, and any additional
number of layers after those three. The appearance of the area layer can be
modified with fixed aesthetics using the area_aes parameter. Fixed aesthetics
for the data layer can be added as additional parameters.
}
\section{Predefined layers}{
NA
}

\concept{layer_area_data}
