% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{layer_area_mask}
\alias{layer_area_mask}
\title{Get a mask for an area}
\usage{
layer_area_mask(
  area = NULL,
  bbox = NULL,
  diag_ratio = NULL,
  dist = NULL,
  asp = NULL,
  crs = NULL,
  ...
)
}
\arguments{
\item{area}{sf object. If multiple areas are provided, they are unioned into a single sf object using \code{\link[sf]{st_union()}}}

\item{bbox}{bbox that optionally defines edge of the mask}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").}

\item{crs}{Selected CRS for returned mask.}
}
\value{
sf object with an outer edge matching the bbox or adjusted bbox with the area cut out.
}
\description{
Returns a mask for an area or areas as an sf object.
Used by the \code{map_area_with_snapbox} function.
}
