% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_area_mask.R
\name{layer_area_mask}
\alias{layer_area_mask}
\title{Add an area mask to a ggplot2 map as a layer}
\usage{
layer_area_mask(
  area = NULL,
  diag_ratio = NULL,
  dist = NULL,
  asp = NULL,
  crs = 2804,
  mask_bbox = NULL,
  ...
)
}
\arguments{
\item{area}{\code{sf} object. Required. If multiple areas are provided, the
areas are combined into a single geometry with \code{\link[sf]{st_union}}}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crs}{Coordinate reference system of bounding box to return}

\item{mask_bbox}{\code{bbox} object to define the edge of the mask.
\code{diag_ratio}, \code{dist}, and \code{asp} parameters are ignored if a
\code{mask_bbox} is provided.}

\item{...}{Additional parameters to pass to \code{\link[ggplot2]{geom_sf}}}
}
\value{
\code{\link[ggplot2]{geom_sf}} function.
}
\description{
Returns a mask for an area or areas as an  \code{sf} object. This function
can be called by any function that uses the logical \code{mask} parameter.
}
