% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_area_streets.R
\name{layer_area_streets}
\alias{layer_area_streets}
\title{Add a layer to a gpplot2 map with area streets, street names, or both}
\usage{
layer_area_streets(
  area = NULL,
  street_type = NULL,
  sha_class = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  msa = FALSE,
  show_streets = TRUE,
  show_names = FALSE,
  name_location = NULL,
  edge_dist = 10,
  ...
)
}
\arguments{
\item{area}{sf object. Returns streets within this area (after adjustment by
dist, diag_ratio, and asp parameters)}

\item{street_type}{selected street subtypes to include. By default, the returned data includes all subtypes except alleys ("STRALY").
Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN")}

\item{sha_class}{selected SHA classifications to include.
"all" selects all streets with an assigned SHA classification (around one-quarter of all street segments).
Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{Logical. Default \code{FALSE}. Trim streets to area using `sf::st_intersection()`.}

\item{msa}{Logical. Default \code{FALSE}. Get streets from cached `baltimore_msa_streets.gpkg` file using `cachedata` parameter of `get_area_data` function.}

\item{show_streets}{Logical. Default TRUE. If FALSE, hides street center
lines.}

\item{show_names}{Logical. Default FALSE. If TRUE, shows street names.}

\item{name_location}{Options include c("area", "edge", "top", "left",
"bottom", "right", "topleft", "topright", "bottomleft", "bottomright").
Defaults to NULL.}

\item{edge_dist}{Distance buffer to use for placing street names.}

\item{...}{Other parameters to pass along to `ggplot2::geom_sf()` that maps
the streets.}
}
\description{
Add a layer to a gpplot2 map with area streets, street names, or both.
}
