% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legislative_districts}
\alias{legislative_districts}
\title{Baltimore City State Legislative Districts}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{name}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{...}
}
\usage{
legislative_districts
}
\description{
...
}
\keyword{datasets}
