% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legislative_districts}
\alias{legislative_districts}
\title{Maryland Legislative Districts for Baltimore City}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{\code{name}}{District name}
\item{\code{id}}{District number}
\item{\code{label}}{District label}
\item{\code{geometry}}{Multipolygon data with district boundaries}
}
}
\source{
\url{https://geodata.md.gov/imap/rest/services/Boundaries/MD_ElectionBoundaries/FeatureServer/1}
}
\usage{
legislative_districts
}
\description{
...
}
\keyword{datasets}
