% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_in_city.R
\name{map_area_in_city}
\alias{map_area_in_city}
\title{Maps an area in the context of the city}
\usage{
map_area_in_city(area, area_label = NULL, map_title = NULL)
}
\arguments{
\item{area}{sf object with a 'name' column. Required.}

\item{area_label}{area label to replace area name. Optional.}

\item{map_title}{custom map title to replace default title. Optional.}
}
\description{
Map showing the location of an area within the city.
}
\examples{

\dontrun{
## Area with a defined label
district2 <- get_area(
  type = "council district",
  area_id = "2"
)

map_area_in_city(
  area = district2,
  area_label = "Council District 2"
)
}

\dontrun{
## Multiple areas in a single map
selected_se_neighborhoods <- get_area(
  type = "neighborhood",
  area_name = c("Upper Fells Point", "Fells Point", "Canton")
)

map_area_in_city(area = selected_se_neighborhoods)
}

\dontrun{
## Area with a defined map title
canton_industrial <- get_area(
  type = "neighborhood",
  area_name = "Canton Industrial Area"
)

map_area_in_city(
  area = canton_industrial,
  map_title = "The Canton Industrial Area is the city's largest neighborhood"
)
}

}
