% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area.R
\name{map_area_in_city}
\alias{map_area_in_city}
\title{Maps an area in the context of the city}
\usage{
map_area_in_city(area, area_label = NULL)
}
\arguments{
\item{area}{sf object with a 'name' column. Required.}

\item{area_label}{area label to replace area name. Optional.}
}
\description{
Map showing the location of an area within the city.
}
\examples{

\dontrun{
## Area with a defined label
district2 <- get_area(
  type = "council district",
  area_id = "2"
)

map_area_in_city(
  area = district2,
  area_label = "Baltimore's Second Council District"
)
}

\dontrun{
## Multiple areas in a single map
selected_se_neighborhoods <- get_area(
  type = "neighborhood",
  area_name = c("Upper Fells Point", "Fells Point", "Canton")
)

map_area_in_city(area = selected_se_neighborhoods,
                 area_label = "Southeast Baltimore neighborhoods")
}

\dontrun{
## Area with a defined map title
canton_industrial <- get_area(
  type = "neighborhood",
  area_name = "Canton Industrial Area"
)

map_area_in_city(area = canton_industrial)
}

}
