% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta_services.R
\name{map_area_mta_services}
\alias{map_area_mta_services}
\title{Map MTA services}
\usage{
map_area_mta_services(area, mta_services = "bus_lines")
}
\arguments{
\item{area}{sf object. Required.}

\item{mta_services}{Character vector. Default is "bus_lines"}
}
\description{
Map MTA services. MTA bus lines are currently the only supported service.
}
