% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\name{map_area_parks}
\alias{map_area_parks}
\title{Map area parks and open spaces}
\usage{
map_area_parks(
  area,
  diag_ratio = 0.125,
  type = c("parks", "vacant lots"),
  label = c("parks")
)
}
\arguments{
\item{area}{sf object. Required.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Passed to \code{\link{get_buffered_area}}.}

\item{type}{layers to show on map ("parks" or "vacant lots"). Defaults to both.}

\item{label}{layers to label. Only "parks" is supported. Use any other value to exclude labels.}
}
\description{
Return a ggplot map showing parks in and around a selected area.
}
