% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_parks.R
\name{map_area_parks}
\alias{map_area_parks}
\title{Map area parks and open spaces}
\usage{
map_area_parks(
  area,
  type = c("parks", "vacant lots"),
  label = c("parks"),
  dist = NULL,
  diag_ratio = 0.125,
  asp = NULL
)
}
\arguments{
\item{area}{sf object. Required.}

\item{type}{layers to show on map ("parks" or "vacant lots"). Defaults to both.}

\item{label}{layers to label. Only "parks" is supported. Use any other value to exclude labels.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}
}
\description{
Return a ggplot map showing parks in and around a selected area.
}
