% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_property.R
\name{map_area_property}
\alias{map_area_property}
\title{Real property or parcel data is from the Maryland State Department of
Assessment and Taxation and may include outdated or inaccurate information.}
\usage{
map_area_property(
  area,
  property = c("improved", "vacant", "principal residence", "use", "building type",
    "value"),
  dist = NULL,
  diag_ratio = 0.1,
  asp = NULL,
  trim = FALSE,
  show_mask = FALSE
)
}
\arguments{
\item{area}{Simple features object. Function currently supports only a single
area at a time.}

\item{property}{Real property variable to map. Options include c("improved",
"vacant", "principal residence", "value"). Currently supports only one
variable at a time.}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{trim}{If TRUE, data trimmed to area with
\code{\link[sf:geos_binary_ops]{sf::st_intersection()}}. This option is not supported for any
adjusted areas that use the \code{dist}, \code{diag_ratio}, or \code{asp} parameters.
Default \code{FALSE}.}

\item{show_mask}{If \code{TRUE}, apply a white, 0.6 alpha mask over property
located outside the provided area. Default \code{FALSE.}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
