% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_with_snapbox.R
\name{map_area_with_snapbox}
\alias{map_area_with_snapbox}
\title{Map area using the snapbox package}
\usage{
map_area_with_snapbox(
  area,
  map_style = snapbox::mapbox_satellite_streets(),
  diag_ratio = 0.125,
  dist = NULL,
  asp = NULL,
  mask = TRUE
)
}
\arguments{
\item{area}{Required sf object with a 'name' column.}

\item{map_style}{Required. \code{\link{stylebox}} function referencing mapbox map styles. Default is \code{\link[stylebox]{mapbox_satellite_streets()}}}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{dist}{buffer distance in meters. Optional.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{mask}{Logical. Default TRUE. If TRUE, apply a transparent (alpha = 0.4) white mask over the Mapbox map outside the area. Uses the layer_area_mask function.}
}
\description{
Map an area or areas using the \code{\link{snapbox}} package.
}
