% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_zoning.R
\name{map_area_zoning}
\alias{map_area_zoning}
\title{Map zoning for an area}
\usage{
map_area_zoning(
  area,
  category = c("all", "residential", "commercial", "industrial"),
  diag_ratio = 0.125
)
}
\arguments{
\item{area}{Required \code{sf} object with a 'name' column.}

\item{category}{Zoning category to return. "all", "residential", "commercial", "industrial"}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}
}
\description{
Map zoning/zoning overlay codes for an area within the city. The 2017 zoning
data does not include any exemptions granted by the BMZA (Board of Municipal
Zoning Appeals).
}
\examples{
\dontrun{
## Map zoning code for Bayview neighborhood
bayview <- get_area(type = "neighborhood", area_name = "Bayview")
map_zoning(area = bayview)
}
}
