% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_area_zoning.R
\name{map_area_zoning}
\alias{map_area_zoning}
\title{Map zoning for an area (not working)}
\usage{
map_area_zoning(
  area,
  category = c("all", "residential", "commercial", "industrial"),
  diag_ratio = 0.125,
  asp = NULL,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804)
)
}
\arguments{
\item{area}{sf, sfc, or bbox object. If multiple areas are provided, they are
unioned into a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}.}

\item{category}{Zoning category to return. "all", "residential", "commercial", "industrial"}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's
bounding box. Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3").}

\item{crs}{Coordinate Reference System (CRS) to use for the returned data.
The CRS of the provided data and bounding box or area must match one
another but are not required to match the CRS provided by this parameter.}
}
\description{
Map zoning/zoning overlay codes for an area within the city. The 2017 zoning
data does not include any exemptions granted by the BMZA (Board of Municipal
Zoning Appeals).
}
