% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_decade_built.R
\name{map_decade_built}
\alias{map_decade_built}
\title{Map real property data to show decade built}
\usage{
map_decade_built(neighborhood_label, neighborhood_color = "gray20")
}
\arguments{
\item{neighborhood_label}{Required. Name of the neighborhood in sentence case}

\item{neighborhood_color}{Optional. Color of the neighborhood boundary line}
}
\description{
Map showing parcels described by decade built.
Parcel data is from the Maryland State Department of Assessment and Taxation and may include errors.
}
