% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_property.R
\name{map_decade_built}
\alias{map_decade_built}
\title{Map real property data to show decade built}
\usage{
map_decade_built(area = NULL)
}
\arguments{
\item{area}{sf class tibble. Object must include a name column.}
}
\description{
Map showing parcels color coded by the decade of the year the primary structure was built. If the area sf tibble includes multiple areas, a separate map is created for each area provided.
Parcel data is from the Maryland State Department of Assessment and Taxation and may include errors.
}
