% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tenure.R
\name{map_tenure}
\alias{map_tenure}
\title{Map real property data to show tenure for neighborhood}
\usage{
map_tenure(area = NULL, area_type = NULL, area_name = NULL, area_label = NULL)
}
\arguments{
\item{area}{sf class tibble. Object must include a name column.}

\item{area_type}{Character vector. Length 1 character vector passed to \code{get_area} function if area is not provided.}

\item{area_name}{Character vector. Passed to \code{get_area} function if area is not provided.}

\item{area_label}{Character vector. Must}
}
\description{
Map showing parcels described as owner occupied, non-owner occupied, vacant, and unimproved.
Parcel data is from the Maryland State Department of Assessment and Taxation and may include errors.
}
