% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_property.R
\name{map_tenure}
\alias{map_tenure}
\title{Map real property data to show tenure for neighborhood}
\usage{
map_tenure(area = NULL)
}
\arguments{
\item{area}{sf class tibble. Object must include a name column.}
}
\description{
Map showing parcels described as principal residence, non-principal residence, vacant, and unimproved properties.
Parcel data is from the Maryland State Department of Assessment and Taxation.
}
