% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoning.R
\name{map_zoning}
\alias{map_zoning}
\title{Map zoning for a local area}
\usage{
map_zoning(neighborhood_label, neighborhood_color = "gray20")
}
\arguments{
\item{neighborhood_label}{Name of the neighborhood in sentence case.}

\item{neighborhood_color}{Color of the neighborhood boundary line}
}
\description{
Map showing the zoning codes for a neighborhood within the city.
The 2017 zoning data does not include any exemptions grants by the BMZA (Board of Municipal Zoning Appeals).
}
