% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbaltimore-package.R
\docType{package}
\name{mapbaltimore-package}
\alias{mapbaltimore}
\alias{mapbaltimore-package}
\title{mapbaltimore: Make maps for Baltimore City with open data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides data from the Baltimore City, the state of Maryland, and other sources, functions to access additional data, and function to create and modify simple maps of Baltimore neighborhoods using sf and ggplot2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/mapbaltimore}
  \item \url{https://elipousson.github.io/mapbaltimore/}
  \item Report bugs at \url{https://github.com/elipousson/mapbaltimore/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
