% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_bus_lines}
\alias{mta_bus_lines}
\title{Maryland Transit Administration Summer 2020 Bus Routes}
\format{
A data frame with 103 rows and 4 variables:
\describe{
  \item{route_name}{Name of the bus route}
  \item{route_type}{Type of route, CityLink, LocalLink and Commuter Bus}
  \item{route_number}{Unique route number or color identifier}
  \item{geometry}{multilinestring with the route path}
  ...
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/maryland-transit-mta-bus-lines-1}
}
\usage{
mta_bus_lines
}
\description{
Maryland Department of Transportation's Maryland Transit Administration
Summer 2020 Bus Routes including CityLink, LocalLink and Commuter Bus.
The data reflect bus route changes as of July 12, 2020.
}
\keyword{datasets}
