% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_bus_lines}
\alias{mta_bus_lines}
\title{Maryland Transit Administration (MTA) Bus Routes}
\format{
A data frame with 103 rows and 4 variables:
\describe{
  \item{\code{route_name}}{Name of the bus route}
  \item{\code{route_type}}{Type of route, CityLink, LocalLink and Commuter Bus}
  \item{\code{route_number}}{Unique route number or color identifier}
  \item{\code{frequent}}{Logical indicator of route inclusion in MTA BaltimoreLink's Frequent Transit Network.}
  \item{\code{geometry}}{Multilinestring with the bus route path}
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/maryland-transit-mta-bus-lines-1}
}
\usage{
mta_bus_lines
}
\description{
Maryland Department of Transportation's Maryland Transit Administration
Summer 2020 Bus Routes including CityLink, LocalLink and Commuter Bus.
The data reflect bus route changes as of July 12, 2020.
}
\keyword{datasets}
