% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_bus_stops}
\alias{mta_bus_stops}
\title{Maryland Transit Administration (MTA) Bus Stops (2022)}
\format{
A data frame with 4426 rows and 11 variables:
\describe{
\item{\code{stop_id}}{Stop identification number}
\item{\code{stop_name}}{Stop name}
\item{\code{rider_on}}{Average daily weekday count of riders boarding transit at stop}
\item{\code{rider_off}}{Average daily weekday count of riders alighting transit at stop}
\item{\code{rider_total}}{Average daily weekday count of total riders served at stop}
\item{\code{stop_ridership_rank}}{Stop rank for ridership}
\item{\code{routes_served}}{Routes served at stop}
\item{\code{mode}}{Mode served at stop}
\item{\code{shelter}}{Logical indicator of bus shelter availability}
\item{\code{county}}{County where stop is located}
\item{\code{direction}}{Route direction}
\item{\code{stop_location}}{Stop location}
\item{\code{frequent}}{Indicator for stop serving frequent transit network}
\item{\code{geometry}}{POINT stop location geometry}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/maryland-transit-mta-bus-stops-1}{Maryland Transit - MTA Bus Stops (MD iMap)}
}
\usage{
mta_bus_stops
}
\description{
Maryland Department of Transportation's Maryland Transit
Administration Bus Stops including CityLink, LocalLink, Express BusLink,
and Commuter Bus. This data is based on the Summer 2022 schedule and
reflects bus stop changes as of June 19, 2022. Ridership data is based on
Automatic Passenger Counting (APC) system average daily weekday bus stop
ridership (boarding, alighting, and total) from the Winter 2022 period and
does not exclude outliers. For full details of service change visit:
\url{https://www.mta.maryland.gov/servicechanges/summer2022}
}
\keyword{datasets}
