% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_light_rail_lines}
\alias{mta_light_rail_lines}
\title{Maryland Transit Administration (MTA) Light RailLink Stations}
\format{
A data frame with 84 rows and 8 variables:
\describe{
  \item{\code{id}}{integer COLUMN_DESCRIPTION}
  \item{\code{rail_name}}{character COLUMN_DESCRIPTION}
  \item{\code{mode}}{character COLUMN_DESCRIPTION}
  \item{\code{tunnel}}{character COLUMN_DESCRIPTION}
  \item{\code{direction}}{character COLUMN_DESCRIPTION}
  \item{\code{miles}}{double COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
  \item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/maryland-transit-light-rail-lines/}{Maryland Transit - Light Rail Lines (MD iMap)}
}
\usage{
mta_light_rail_lines
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
