% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_light_rail_stations}
\alias{mta_light_rail_stations}
\title{Maryland Transit Administration (MTA) Light RailLink Stations}
\format{
A data frame with 33 rows and 11 variables:
\describe{
\item{\code{id}}{Feature ID}
\item{\code{name}}{Station name}
\item{\code{address}}{Station address}
\item{\code{city}}{City}
\item{\code{state}}{State}
\item{\code{zipcode}}{Zipcode}
\item{\code{mode}}{Facility mode (Light Rail)}
\item{\code{avg_wkdy}}{Average weekday passengers}
\item{\code{avg_wknd}}{Average weekend passengers}
\item{\code{facility_type}}{Facility type}
\item{\code{geometry}}{POINT geometry for station locations}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/maryland-transit-light-raillink-stations/}{Maryland Transit - Light RailLink Stations (MD iMap)}
}
\usage{
mta_light_rail_stations
}
\description{
Locations for stations on the Baltimore Light RailLink (Baltimore Light Rail)
line operated by the Maryland Transit Administration.
}
\keyword{datasets}
