% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_marc_lines}
\alias{mta_marc_lines}
\title{Maryland Transit Administration (MTA) MARC Train Lines}
\format{
A data frame with 162 rows and 8 variables:
\describe{
\item{\code{id}}{Feature ID}
\item{\code{rail_name}}{Rail line name}
\item{\code{mode}}{Facility mode and line name (MARC)}
\item{\code{tunnel}}{Tunnel indicator}
\item{\code{direction}}{Travel direction}
\item{\code{miles}}{Section mileage}
\item{\code{status}}{Section status}
\item{\code{geometry}}{LINESTRING geometry with rail lines}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/de0efbe9f8884ac5aa69864b6b3ff633_10/}{Maryland Transit - MARC Train Lines (MD iMap)}
}
\usage{
mta_marc_lines
}
\description{
MARC (Maryland Area Regional Commuter) Rail system lines operated by the
Maryland Transit Administration.
}
\keyword{datasets}
