% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_marc_stations}
\alias{mta_marc_stations}
\title{Maryland Transit Administration (MTA) MARC Train Stations}
\format{
A data frame with 44 rows and 12 variables:
\describe{
\item{\code{id}}{Feature ID}
\item{\code{name}}{Station name}
\item{\code{address}}{Station address}
\item{\code{city}}{City}
\item{\code{state}}{State}
\item{\code{zipcode}}{Zipcode}
\item{\code{line_name}}{Line name}
\item{\code{mode}}{Facility mode and line name (MARC)}
\item{\code{avg_wkdy}}{Average weekday passengers}
\item{\code{avg_wknd}}{Average weekend passengers}
\item{\code{facility_type}}{Facility type (Station)}
\item{\code{geometry}}{POINT geometry with station locations}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/maryland-transit-marc-trains-stations/}{Maryland Transit - MARC Trains Stations (MD iMap)}
}
\usage{
mta_marc_stations
}
\description{
Locations of MARC (Maryland Area Regional Commuter) Rail stations operated by the
Maryland Transit Administration.
}
\keyword{datasets}
