% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_subway_lines}
\alias{mta_subway_lines}
\title{Maryland Transit Administration (MTA) SubwayLink Metro Lines}
\format{
A data frame with 34 rows and 8 variables:
\describe{
\item{\code{id}}{Feature id number as integer}
\item{\code{rail_name}}{Subway line name (Metro Line)}
\item{\code{mode}}{Travel mode (Metro)}
\item{\code{tunnel}}{Section tunnel indicator}
\item{\code{direction}}{Travel direction}
\item{\code{miles}}{Section mileage}
\item{\code{status}}{Section status}
\item{\code{geometry}}{MULTILINESTRING geometry for lines}
}
}
\source{
\href{https://geodata.md.gov/imap/rest/services/Transportation/MD_Transit/FeatureServer/5}{Baltimore Metro Subway Line}
}
\usage{
mta_subway_lines
}
\description{
Route of MTA SubwayLink Metro Line.
}
\keyword{datasets}
