% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_subway_stations}
\alias{mta_subway_stations}
\title{Maryland Transit Administration (MTA) SubwayLink Metro Stations}
\format{
A data frame with 14 rows and 10 variables:
\describe{
  \item{\code{id}}{integer COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{address}}{character COLUMN_DESCRIPTION}
  \item{\code{city}}{character COLUMN_DESCRIPTION}
  \item{\code{state}}{character COLUMN_DESCRIPTION}
  \item{\code{mode}}{character COLUMN_DESCRIPTION}
  \item{\code{avg_wkdy}}{integer COLUMN_DESCRIPTION}
  \item{\code{avg_wknd}}{integer COLUMN_DESCRIPTION}
  \item{\code{facility_type}}{character COLUMN_DESCRIPTION}
  \item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\usage{
mta_subway_stations
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
