% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{named_intersections}
\alias{named_intersections}
\title{Baltimore City street intersection names}
\format{
A data frame with 11506 rows and 3 variables:
\describe{
  \item{\code{id}}{Intersection identifier matching id in `edge_of_pavement` data}
  \item{\code{name}}{Intersection name}
  \item{\code{geometry}}{Points with center of intersections}
}
}
\usage{
named_intersections
}
\description{
Index of Baltimore City intersections with names from streets
  within 20 meters of the intersection boundaries.
}
\keyword{datasets}
