% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhoods}
\alias{neighborhoods}
\title{Baltimore City Neighborhood Boundaries}
\format{
A data frame with 278 rows and 2 variables:
\describe{
  \item{label}{Label with the name of the neighborhood}
  \item{neighborhood}{Neighborhood name capitalized}
  \item{geometry}{Multipolygons with boundary of each neighborhood}
  ...
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/fc5d183b20a145009eae8f8b171eeb0d_0}
}
\usage{
neighborhoods
}
\description{
Baltimore City Neighborhoods or Neighborhood Statistical Areas
}
\keyword{datasets}
