% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhoods}
\alias{neighborhoods}
\title{Neighborhood boundaries for Baltimore City}
\format{
A data frame with 278 rows and 2 variables:
\describe{
  \item{\code{name}}{Neighborhood name}
  \item{\code{name}}{Type of area, with options including residential, industrial area, park, and business park)}
  \item{\code{acres}}{Area of the neighborhood (acres)}
  \item{\code{geometry}}{Multipolygons with neighborhood boundary}
}
}
\source{
\href{https://data.imap.maryland.gov/datasets/fc5d183b20a145009eae8f8b171eeb0d_0}{Maryland Baltimore City Neighborhoods (MD iMap)}
}
\usage{
neighborhoods
}
\description{
Baltimore City Neighborhoods or Neighborhood Statistical Areas
}
\keyword{datasets}
