% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parks}
\alias{parks}
\title{Baltimore City Parks}
\format{
A data frame with 297 rows and 9 variables:
\describe{
\item{\code{name}}{Park name}
\item{\code{id}}{Identification number from city data}
\item{\code{address}}{Primary street address}
\item{\code{name_alt}}{Alternate name}
\item{\code{operator}}{Park operator, Baltimore City Department of Recreation and Parks (BCRP) or other}
\item{\code{park_district}}{Park maintenance district for BCRP}
\item{\code{acres}}{Area of the park property (acres)}
\item{\code{osm_id}}{OpenStreetMap ID (node, way, or relation)}
\item{\code{geometry}}{Multipolygon with park edges}
}
}
\source{
\url{https://services1.arcgis.com/UWYHeuuJISiGmgXx/ArcGIS/rest/services/Baltimore_City_Recreation_and_Parks/FeatureServer/2}
}
\usage{
parks
}
\description{
Spatial data for parks and public recreation centers in Baltimore City from
the Baltimore City Department of Recreation and Parks. A few names have been
updated to use common names or recent new official names so the package
version may not match the city data in all cases. The parks have been matched
to corresponding entities on OpenStreetMap indicated by the osm_id column.
}
\keyword{datasets}
