% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parks}
\alias{parks}
\title{Baltimore City Parks}
\format{
A data frame with 297 rows and 6 variables:
\describe{
  \item{name}{Park name}
  \item{id}{Identification number from city GIS data}
  \item{address}{Primary street address}
  \item{name_alt}{Alternate name}
  \item{operator}{Park operator, Baltimore City Department of Recreation and Parks or other}
  \item{area}{Area of the park property in acres}
  \item{geometry}{Multipolygon with edges of parks}
  ...
}
}
\source{
\url{https://services1.arcgis.com/UWYHeuuJISiGmgXx/ArcGIS/rest/services/Baltimore_City_Recreation_and_Parks/FeatureServer/2}
}
\usage{
parks
}
\description{
Spatial data for parks in Baltimore City from the Baltimore City Department of Recreation and Parks.
}
\keyword{datasets}
