% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planning_districts}
\alias{planning_districts}
\title{Baltimore City Planning Districts}
\format{
A data frame with 11 rows and 4 variables:
\describe{
  \item{\code{id}}{Planning district area identifier}
  \item{\code{name}}{Full name of the planning district}
  \item{\code{abb}}{Planning district area abbreviation}
  \item{\code{geometry}}{Multipolygon boundary of the planning district}
}
}
\source{
\url{https://geodata.baltimorecity.gov/egis/rest/services/CityView/PlanningDistricts/MapServer/0}
}
\usage{
planning_districts
}
\description{
Administrative boundaries set by the Baltimore City Department of Planning.
}
\keyword{datasets}
