% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police_districts}
\alias{police_districts}
\title{Baltimore City Police Districts}
\format{
A data frame with 9 rows and 3 variables:
\describe{
  \item{number}{Police district number}
  \item{name}{Police district name}
  \item{geometry}{Multipolygons with boundary of each district}
  ...
}
}
\source{
\url{...}
}
\usage{
police_districts
}
\description{
Baltimore City Police Districts
}
\keyword{datasets}
