% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real_property}
\alias{real_property}
\title{Baltimore City Real Property Data}
\format{
A data frame with 238,340 rows and 82 variables:
\describe{
  \item{objectid}{...}
  \item{pin}{...}
  \item{pinrelate}{...}
  \item{blocklot}{...}
  \item{block}{...}
  \item{lot}{...}
  \item{ward}{...}
  \item{section}{...}
  \item{assessor}{...}
  \item{taxbase}{...}
  \item{bfcvland}{...}
  \item{bfcvimpr}{...}
  \item{landexmp}{...}
  \item{imprexmp}{...}
  \item{citycred}{...}
  \item{statcred}{...}
  \item{ccredamt}{...}
  \item{scredamt}{...}
  \item{permhome}{...}
  \item{assesgrp}{...}
  \item{lot_size}{...}
  \item{no_imprv}{...}
  \item{currland}{...}
  \item{currimpr}{...}
  \item{exmpland}{...}
  \item{exmpimpr}{...}
  \item{fullcash}{...}
  \item{exmptype}{...}
  \item{exmpcode}{...}
  \item{usegroup}{...}
  \item{zonecode}{...}
  \item{sdatcode}{...}
  \item{artaxbas}{...}
  \item{distswch}{...}
  \item{dist_id}{...}
  \item{statetax}{...}
  \item{city_tax}{...}
  \item{ar_owner}{...}
  \item{deedbook}{...}
  \item{deedpage}{...}
  \item{saledate}{...}
  \item{owner_abbr}{...}
  \item{owner_1}{...}
  \item{owner_2}{...}
  \item{owner_3}{...}
  \item{fulladdr}{...}
  \item{stdirpre}{...}
  \item{st_name}{...}
  \item{st_type}{...}
  \item{bldg_no}{...}
  \item{fraction}{...}
  \item{unit_num}{...}
  \item{span_num}{...}
  \item{spanfrac}{...}
  \item{zip_code}{...}
  \item{extd_zip}{...}
  \item{dhcduse1}{...}
  \item{dhcduse2}{...}
  \item{dhcduse3}{...}
  \item{dhcduse4}{...}
  \item{dwelunit}{...}
  \item{eff_unit}{...}
  \item{roomunit}{...}
  \item{rpdeltag}{...}
  \item{respagcy}{...}
  \item{salepric}{...}
  \item{propdesc}{...}
  \item{neighbor}{...}
  \item{srvccntr}{...}
  \item{year_build}{...}
  \item{structarea}{...}
  \item{ldate}{...}
  \item{ownmde}{...}
  \item{grndrent}{...}
  \item{subtype_geodb}{...}
  \item{sdatlink}{...}
  \item{blockplat}{...}
  \item{mailtoadd}{...}
  \item{vacind}{...}
  \item{shape_st_area}{...}
  \item{shape_st_length}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{https://gis-baltimore.opendata.arcgis.com/datasets/real-property}
}
\usage{
real_property
}
\description{
This dataset represents the City of Baltimore parcel boundaries, with
ownership, address, vaulation and other property information.
This data was downloaded on October 6, 2020.
}
\keyword{datasets}
