% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mapbaltimore.R
\name{scale_color_mapbaltimore}
\alias{scale_color_mapbaltimore}
\alias{scale_fill_mapbaltimore}
\title{Scales for Baltimore data}
\usage{
scale_color_mapbaltimore(palette = NULL)

scale_fill_mapbaltimore(palette = NULL)
}
\arguments{
\item{palette}{"mta_bus" is only currently supported option, Default: NULL}
}
\description{
Scales for Baltimore data
}
\seealso{
\code{\link[ggplot2]{scale_manual}}
}
