% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mapbaltimore.R
\name{scale_mapbaltimore}
\alias{scale_mapbaltimore}
\alias{scale_color_mapbaltimore}
\alias{scale_fill_mapbaltimore}
\title{Scales for Baltimore data}
\usage{
scale_mapbaltimore(
  palette = NULL,
  values = NULL,
  na.value = "grey50",
  aesthetics = c("color", "fill"),
  error_call = caller_env(),
  ...
)

scale_color_mapbaltimore(palette = NULL, na.value = "grey50", ...)

scale_fill_mapbaltimore(palette = NULL, na.value = "grey50", ...)
}
\arguments{
\item{palette}{Options include "mta_bus", "hmt_2017", "hmt_cluster",
"cluster", "hmt_cluster_group", or "cluster_group", Default: \code{NULL}}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}

\item{na.value}{Defaults to "grey50"}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:discrete_scale]{discrete_scale}}
  \describe{
    \item{\code{limits}}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{drop}}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} includes the levels in the factor. Please note that to display
every level in a legend, the layer should use \code{show.legend = TRUE}.}
    \item{\code{na.translate}}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
    \item{\code{name}}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
    \item{\code{labels}}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{guide}}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}
    \item{\code{call}}{The \code{call} used to construct the scale for reporting messages.}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}
}
\description{
Custom palettes for two package datasets: \code{mta_bus_lines} and \code{hmt_2017}
(both for cluster and cluster group).
}
\examples{
\dontrun{
if (interactive()) {
  library(ggplot2)

  ggplot(data = dplyr::filter(mta_bus_lines, frequent)) +
    geom_sf(aes(color = route_abb), alpha = 0.5, size = 2) +
    scale_mapbaltimore(palette = "bus") +
    theme_minimal()

  ggplot(data = hmt_2017) +
    geom_sf(aes(fill = cluster_group, color = cluster_group)) +
    scale_mapbaltimore(palette = "cluster_group") +
    theme_minimal()
}
}

}
