% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_map_limits}
\alias{set_map_limits}
\title{Set map limits to area}
\usage{
set_map_limits(
  area = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crs = 2804
)
}
\arguments{
\item{area}{sf object}

\item{bbox}{bbox object}

\item{dist}{buffer distance in meters. Optional.}

\item{diag_ratio}{ratio to set map extent based diagonal distance of area's bounding box. Default is 0.125 (1/8). Ignored when \code{dist} is provided.}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").}

\item{crs}{Coordinate reference system to use. Default EPSG:2804.}
}
\value{
ggplot2::coord_sf() function to modify a ggplot2 object
}
\description{
Set map limits to area with optional buffer or aspect ratio adjustment
}
\details{
Set limits for a map to the bounding box of an area using coord_sf.
Optionally, adjust the area size using get_buffered_area function
and/or adjust the aspect ratio of the limiting bounding box to match
a set aspect ratio.
}
