% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streets}
\alias{streets}
\title{Baltimore City Street Center lines}
\format{
Simple feature collection with 48,473 features and 34 fields.
\describe{
  \item{objectid_1}{...}
  \item{objectid}{...}
  \item{tag}{...}
  \item{last_org}{...}
  \item{capture_me}{...}
  \item{last_user}{...}
  \item{edit_date}{...}
  \item{type}{...}
  \item{subtype}{...}
  \item{dirpre}{...}
  \item{feanme}{...}
  \item{featype}{...}
  \item{dirsuf}{...}
  \item{fraddl}{...}
  \item{toaddl}{...}
  \item{fraddr}{...}
  \item{toaddr}{...}
  \item{fraddla}{...}
  \item{toaddla}{...}
  \item{fraddra}{...}
  \item{toaddra}{...}
  \item{leftzip}{...}
  \item{rightzip}{...}
  \item{fullname}{...}
  \item{flag}{...}
  \item{comments}{...}
  \item{sha_class}{...}
  \item{shape_leng}{...}
  \item{place}{...}
  \item{blocktext}{...}
  \item{block_num}{...}
  \item{zipcode}{...}
  \item{global_id}{...}
  \item{url}{...}
  \item{geometry}{...}
  ...
}
}
\source{
\url{...}
}
\usage{
streets
}
\description{
...
}
\keyword{datasets}
