% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streets}
\alias{streets}
\title{Baltimore City Street Center lines}
\format{
Simple feature collection with 48,473 features and 23 fields.
\describe{
  \item{\code{type}}{...}
  \item{\code{subtype}}{...}
  \item{\code{subtype_label}}{...}
  \item{\code{dirpre}}{...}
  \item{\code{feanme}}{...}
  \item{\code{featype}}{...}
  \item{\code{dirsuf}}{...}
  \item{\code{fraddl}}{...}
  \item{\code{toaddl}}{...}
  \item{\code{fraddr}}{...}
  \item{\code{toaddr}}{...}
  \item{\code{fraddla}}{...}
  \item{\code{toaddla}}{...}
  \item{\code{fraddra}}{...}
  \item{\code{toaddra}}{...}
  \item{\code{leftzip}}{...}
  \item{\code{rightzip}}{...}
  \item{\code{fullname}}{...}
  \item{\code{sha_class}}{...}
  \item{\code{sha_class_label}}{...}
  \item{\code{blocktext}}{...}
  \item{\code{block_num}}{...}
  \item{\code{geometry}}{...}
}
}
\source{
\url{https://dotgis.baltimorecity.gov/arcgis/rest/services/DOT_Map_Services/DOT_Basemap/MapServer/7}
}
\usage{
streets
}
\description{
...
}
\keyword{datasets}
