% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wards_1797_1918}
\alias{wards_1797_1918}
\title{Baltimore City Ward Boundaries, 1797-1918}
\format{
A data frame with 245 rows and 4 variables:
\describe{
  \item{year}{Earliest effective year of ward boundaries}
  \item{name}{Ward name}
  \item{number}{Ward number}
  \item{geometry}{Multipolygons with the ward boundaries}
  ...
}
}
\source{
\url{https://msa.maryland.gov/bca/wards/index.html}
}
\usage{
wards_1797_1918
}
\description{
Historic ward boundary data from 1797 to 1918. Derived from KML data provided by the Baltimore City Archives.
}
\keyword{datasets}
