% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoning}
\alias{zoning}
\title{Baltimore City Zoning Code}
\format{
A data frame with 2,406 rows and 4 variables:
\describe{
  \item{zoning}{Zoning designation code}
  \item{overlay}{Overlay zone designation}
  \item{label}{Label combining zoning and overlay zoning codes}
  \item{geoms}{Multipolygons for areas with shared zoning}
  ...
}
}
\source{
\url{https://geodata.baltimorecity.gov/egis/rest/services/Planning/Boundaries_and_Plans/MapServer/20}
}
\usage{
zoning
}
\description{
The Baltimore City Zoning Code is administered by the Baltimore City
Department of Housing and Community Development (HCD) Office of the
Zoning Administrator. This office supports the Board of Municipal Zoning Appeals (BMZA).
}
\keyword{datasets}
